### 2021-02-21 #107 @whedon
[PRE REVIEW]: IndeterminateBeam: A Python package for solving 1D indeterminate beams
**Submitting author:** @jessebonanno (<a href="http://orcid.org/0000-0002-4996-6813">Jesse Bonanno</a>)
**Repository:** <a href="https://github.com/JesseBonanno/IndeterminateBeam/" target ="_blank">https://github.com/JesseBonanno/IndeterminateBeam/</a>
**Version:** v2.0.1

@jessebonanno if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
