### 2021-02-21 #107 @labarba
Hi @moorepants — you handled the [Beambending](https://jose.theoj.org/papers/10.21105/jose.00065) submission, which this package extends. Can you take this through review?

One question I have is why create a new package and not add functionality to the existing one? It seems like having so many packages to study statics would not be great for the users...
