### 2021-02-22 ac78739
speed fix and data points
Fixed loading m0 issues for speed. Gave a beam a property for plotting data to allow for switching the number of data points when using the dash application, and added a dash slider to modify this in dash. Fixed add_load function to not add loads with 0 force.
