+++ tests/test_indeterminatebeam_speed.py
@@ -29,9 +29,11 @@ import time
 
 class BeamTestCase(unittest.TestCase):
     def test_speed(self):
+        t1 = time.perf_counter()
         self.store_speeds = []
         self.filename_speeds = input('What is the test name?')
         # case 1
+        print(t1 - time.perf_counter())
         t1 = time.perf_counter()
 
         beam = Beam(5)
@@ -43,19 +45,36 @@ class BeamTestCase(unittest.TestCase):
         # e = DistributedLoad(5,(0,1),45)
         # f = DistributedLoadV(5,(0,1))
         # g = DistributedLoadH(5,(0,1))
-        h = TrapezoidalLoad((0,1),(0,1),45)
+        # h = TrapezoidalLoad((0,1),(0,1),45)
         i = TrapezoidalLoadV((0,1),(0,1))
         j = TrapezoidalLoadH((0,1),(0,1))
 
-        beam.add_loads(a,b,c,d,h,i,j)
+        print(t1 - time.perf_counter())
+        t1 = time.perf_counter()
+
+        beam.add_loads(a,b,c,d)
         beam.add_supports(
             Support(),
             Support(5,(1,1,0))
         )
+
+        print(t1 - time.perf_counter())
+        t1 = time.perf_counter()
+
         beam.analyse()
 
+        print(t1 - time.perf_counter())
+        t1 = time.perf_counter()
+
         fig = beam.plot_beam_external()
+
+        print(t1 - time.perf_counter())
+        t1 = time.perf_counter()
+        
         fig = beam.plot_beam_internal()
+
+        print(t1 - time.perf_counter())
+        t1 = time.perf_counter()
 
         self.store_speeds.append(t1-time.perf_counter())
 
