### 2021-02-22 #107 @moorepants
@JesseBonanno, as a pre-review check can you explain here which of the two submission types this submission fits in? You can find the submission requirements here: https://jose.theoj.org/about. I see that you have created a software library for making calculations regarding beams, but that does not fit our specification for "educational software tools". I also do not see any teaching and learning content for use in course lessons which is needed for the "open-source education materials" submission type.
