### 2021-02-22 #107 @JesseBonanno
@moorepants I have intended this to be an educational software tool that can be used to check the solutions for beams. It can be thought of as an autograder for students solving for beams. I have also included documentation about the theory behind the solutions. As to why i didnt simply update the original package, i am new to coding and was originally playing around. By now i would say there are no lines from the original package in the the IndeterminateBeam code. It is a completely new code, and is a few thousand lines in length vs the original few hundred.
