### 2021-02-23 be77ee7
documentation update
Adding comments and fixing docstrings.
