+++ indeterminatebeam/plotly_drawing_aid.py
@@ -17,6 +17,7 @@ from indeterminatebeam.loading import(
     DistributedLoad,
 )
 
+
 def draw_line(fig, angle, x_sup, length=-20, xoffset=0, yoffset=0,
               color='red', line_width=2, row=None, col=None):
     """Draw an anchored line on a plotly figure.
@@ -152,7 +153,8 @@ def draw_arrowhead(fig, angle, x_sup, length=5, xoffset=0, yoffset=0,
 
 
 def draw_arrow(fig, angle, force, x_sup, xoffset=0, yoffset=0, color='red',
-               line_width=2, arrowhead=5, arrowlength=40, show_values=True, row=None, col=None):
+               line_width=2, arrowhead=5, arrowlength=40, show_values=True,
+               row=None, col=None):
     """Draw an anchored arrow on a plotly figure.
 
     Parameters
@@ -201,7 +203,7 @@ def draw_arrow(fig, angle, force, x_sup, xoffset=0, yoffset=0, color='red',
     # Factor to switch arrow direction based on force sign
     if force > 0:
         d = 1
-    elif force<0:
+    elif force < 0:
         d = -1
     else:
         return fig
@@ -235,15 +237,21 @@ def draw_arrow(fig, angle, force, x_sup, xoffset=0, yoffset=0, color='red',
         # determine start and end of arrow
         x0 = xoffset + x_sup
         y0 = yoffset
-        x1 = (x0 + int(sympify(-arrowlength* d * cos(radians(angle))).evalf(2)))*1.1
-        y1 = (y0 + int(sympify(-arrowlength * d * sin(radians(angle))).evalf(2)))*1.3
-        
+        x1 = (
+            x0
+            + int(sympify(-arrowlength * d * cos(radians(angle))).evalf(2))
+            ) * 1.1
+        y1 = (
+            y0
+            + int(sympify(-arrowlength * d * sin(radians(angle))).evalf(2))
+            ) * 1.3
+
         # make so text doesnt intersect x axis
-        if abs(y1)<5:
+        if abs(y1) < 5:
             if y1 >= 0:
-                y1=10
+                y1 = 10
             else:
-                y1=-10
+                y1 = -10
 
         annotation = dict(
             xref="x", yref="y",
@@ -254,7 +262,7 @@ def draw_arrow(fig, angle, force, x_sup, xoffset=0, yoffset=0, color='red',
             text=force,
             font_color=color,
             showarrow=False,
-            )
+        )
 
         # Append shape to plot or subplot
         if row and col:
@@ -373,8 +381,14 @@ def draw_support_rectangle(fig, x_sup, orientation="up", row=None, col=None):
     return fig
 
 
-def draw_moment(fig, moment, x_sup, color='magenta', show_values=True, row=None,
-                col=None):
+def draw_moment(
+        fig,
+        moment,
+        x_sup,
+        color='magenta',
+        show_values=True,
+        row=None,
+        col=None):
     """Draw a moment (torque) shape (circular arrow) on a plotly figure.
 
     Parameters
@@ -437,7 +451,7 @@ def draw_moment(fig, moment, x_sup, color='magenta', show_values=True, row=None,
             text=moment,
             font_color=color,
             showarrow=False,
-            )
+        )
 
         # Append shape to plot or subplot
         if row and col:
@@ -480,7 +494,7 @@ def draw_force(fig, load, row=None, col=None):
             row=row,
             col=col)
 
-    elif isinstance(load,PointLoad):
+    elif isinstance(load, PointLoad):
         force, x_sup, angle = load.force, load.position, load.angle
 
         fig = draw_arrow(
@@ -502,12 +516,11 @@ def draw_force(fig, load, row=None, col=None):
         elif angle % 90 != 0:
[…]
