+++ indeterminatebeam/data_validation.py
@@ -10,6 +10,7 @@ def assert_positive_number(n, name):
         raise ValueError(
             f"The value for '{name}' should be >= 0, not {n}")
 
+
 def assert_strictly_positive_number(n, name):
     """Assert that an input is a strictly positive number."""
     if type(n) not in [int, float]:
@@ -25,6 +26,7 @@ def assert_number(n, name):
     if type(n) not in [int, float]:
         raise ValueError(
             f"The value for '{name}' should be an integer or a float, not a {type(n)}.")
+
 
 def assert_length(var, num, name):
     """Assert that a tuple or list is of a specific length."""
