+++ indeterminatebeam/loading.py
@@ -6,36 +6,39 @@ from sympy import oo, integrate, SingularityFunction, sympify, cos, sin
 from math import radians
 
 from indeterminatebeam.data_validation import (
-    assert_length, 
-    assert_number, 
-    assert_positive_number, 
+    assert_length,
+    assert_number,
+    assert_positive_number,
     assert_strictly_positive_number
 )
 
+
 class Load:
     """Load class from which all other types of loads inherit."""
+
     def _add_load_functions(self, angle, expr):
         # x and y vectors for force
         force_x = cos(radians(angle))
         force_y = sin(radians(angle))
 
         # self._x0 represents w(x), load as a function of x
-        if abs(round(force_x,8)) > 0:
+        if abs(round(force_x, 8)) > 0:
             self._x0 = force_x * expr
-        else: 
+        else:
             self._x0 = 0
 
         # self._y0 represents w(x), load as a function of x
-        if abs(round(force_y,8)) > 0:
+        if abs(round(force_y, 8)) > 0:
             self._y0 = force_y * expr
-        else: 
+        else:
             self._y0 = 0
 
         # self._x1 represents NF(x), normal force as function of x.
-        self._x1 = integrate(self._x0, x) #NF
+        self._x1 = integrate(self._x0, x)  # NF
 
         # self._y1 represents SF(x), shear force as function of x.
-        self._y1 = integrate(self._y0, x) #SF
+        self._y1 = integrate(self._y0, x)  # SF
+
 
 class PointTorque(Load):
     """Point clockwise torque.
@@ -53,27 +56,28 @@ class PointTorque(Load):
     # 30 kN·m (clockwise) torque at x=4 m
     >>> motor_torque = PointTorque(30, 4)
     """
-    
-    def __init__(self, force = 0, coord=0):
+
+    def __init__(self, force=0, coord=0):
         # Data Validation.
         assert_number(force, 'force')
         assert_positive_number(coord, 'coordinate')
-        
-        #load as a function of x (non-directional)
+
+        # load as a function of x (non-directional)
         expr = force * SingularityFunction(x, coord, -1)
-        
+
         # add load as a function of x (directional)
         # and add integration of this function to object.
         # angle 90 as moment only affects vertical direction.
         angle = 90
         self._add_load_functions(angle, expr)
-    
+
         # self._m0 is moment induced by load about coord 0.
         self._m0 = force
-        
+
         # Assign other inputs to load object
         self.position = coord
         self.force = force
+
 
 class PointLoad(Load):
     """Point load.
@@ -100,8 +104,8 @@ class PointLoad(Load):
     >>> external_force
     PointLoad(force=-30, coord=3, angle=0)
     """
-    
-    def __init__(self, force = 0, coord=0, angle=0):
+
+    def __init__(self, force=0, coord=0, angle=0):
         # Data Validation for inputs
         assert_number(force, 'force')
         assert_positive_number(coord, 'coordinate')
@@ -113,7 +117,7 @@ class PointLoad(Load):
         # add load as a function of x (directional)
         # and add integration of this function to object
         self._add_load_functions(angle, expr)
-        
+
         # self._m0 is moment induced by load about coord 0.
         force_y = sin(radians(angle))
         self._m0 = force * coord * force_y
@@ -122,6 +126,7 @@ class PointLoad(Load):
         self.position = coord
         self.force = force
         self.angle = angle
+
 
 class UDL(Load):
     """Uniformly Distributed Load (UDL).
@@ -144,15 +149,14 @@ class UDL(Load):
     # load of 1 kN/m from 1 <= x <= 4 (vertical)
[…]
