### 2021-02-24 6eb6093
Documentation (imports, indeterminatebeam)
Fixed imports to match PEP8 order. Started ammending comments and docstrings in indeterminatebeam module.
