+++ indeterminatebeam/indeterminatebeam.py
@@ -1,6 +1,5 @@
-"""Main module that contains the main class Beam, and auxiliary classes
-Support, PointLoadH, PointLoadV, DistributedLoadH, and DistributedLoadV,
-PointLoad, PointTorque and TrapezoidalLoad.
+"""Main module that contains the main class for Beam 
+and auxillary class for Support.
 
 Example
 -------
@@ -12,27 +11,26 @@ Example
 >>> beam.analyse()
 >>> beam.plot()
 """
-import sys, os
-sys.path.insert(0, os.path.abspath('../'))
 
+# Standard Library Imports
 from collections import namedtuple
 from copy import deepcopy
+from math import radians
+
+# Third Party Imports
 import numpy as np
-import os
 from sympy import (integrate, lambdify, Piecewise, sympify, symbols, 
                    linsolve, sin, cos, oo, SingularityFunction)
 from sympy.abc import x
-from math import radians
+from plotly.subplots import make_subplots
+import plotly.graph_objects as go
 
-from indeterminatebeam.data_validation import (assert_number, assert_positive_number,
-                             assert_strictly_positive_number, assert_length)
-from indeterminatebeam.plotly_drawing_aid import (
-    draw_line, draw_arrowhead, draw_arrow, draw_support_triangle,
-    draw_support_rectangle, draw_moment, draw_force, draw_load_hoverlabel,
-    draw_reaction_hoverlabel, draw_support_hoverlabel, draw_support_rollers,
-    draw_support_spring, draw_support
-    )
-
+# Local Application Imports
+from indeterminatebeam.data_validation import (
+    assert_number,
+    assert_positive_number,
+    assert_strictly_positive_number, assert_length
+)
 from indeterminatebeam.loading import (
     PointTorque,
     PointLoad,
@@ -40,11 +38,23 @@ from indeterminatebeam.loading import (
     TrapezoidalLoad,
     DistributedLoad,
 )
+from indeterminatebeam.plotly_drawing_aid import (
+    draw_line, 
+    draw_arrowhead, 
+    draw_arrow, 
+    draw_support_triangle,
+    draw_support_rectangle, 
+    draw_moment, 
+    draw_force, 
+    draw_load_hoverlabel,
+    draw_reaction_hoverlabel, 
+    draw_support_hoverlabel, 
+    draw_support_rollers,
+    draw_support_spring, 
+    draw_support
+)
 
-from plotly.subplots import make_subplots
-import plotly.graph_objects as go
 
-import time
 
 class Support:
     """
@@ -66,9 +76,6 @@ class Support:
             Degrees of freedom that are completely fixed on a beam for
             movement in x, y and bending, 1 represents fixed and 0
             represents free or spring  (default (1,1,1))
-        _id : positive number
-            id assigned when support associated with Beam object, to
-            help remove supports.
 
     Examples
     --------
@@ -103,30 +110,35 @@ class Support:
             stiffness of y support (kN/mm), if set will overide the
             value placed in the fixed tuple. (default = None)
         """
-        # input validation
+        # validate coordinate
         assert_positive_number(coord, 'coordinate')
-
+        
+        # validate support stiffness if assigned
         if kx:
             assert_positive_number(kx, 'kx')
         if ky:
             assert_positive_number(ky, 'ky')
 
+        # validate fixed tuple elements
         for a in fixed:
             if a not in [0, 1]:
                 raise ValueError(
                     "The provided fixed parameter, must be a tuple of \
                     booleans of length 3"
                 )
+
+        # validate fixed tuple length
         if len(fixed) != 3:
             raise ValueError(
                 "The provided fixed parameter, must be a tuple of \
                 booleans of length 3"
             )
 
[…]
