### 2021-02-24 12c3950
Update indeterminatebeam.py
