+++ indeterminatebeam/indeterminatebeam.py
@@ -14,7 +14,6 @@ Example
 
 # Standard Library Imports
 from collections import namedtuple
-from copy import deepcopy
 from math import radians
 
 # Third Party Imports
@@ -201,9 +200,6 @@ class Beam:
     _deflection_equation: sympy piecewise function
         A sympy function representing the tangential deflection as 
         a function of x.
-    _axial_deflection: sympy piecewise function
-        A sympy function representing the axial defelection as a
-        function of x.
 
     _reactions: dictionary of lists
         A dictionary with keys for support positions. Each key is
@@ -267,7 +263,6 @@ class Beam:
         self._shear_forces = 0
         self._bending_moments = 0
         self._deflection_equation = 0
-        self._axial_deflection = 0
 
         self._reactions = {}
         self._DATA_POINTS = 200
@@ -446,7 +441,7 @@ class Beam:
         # sympy variable and entry in unknowns dictionary.
         for a in self._supports:
             if a._stiffness[0] != 0:
-                unknowns['x'].append()
+                unknowns['x'].append(
                     {
                         'position':a._position,
                         'stiffness':a._stiffness[0],
@@ -650,23 +645,6 @@ class Beam:
         self._deflection_equation = self.sympy_expr_to_piecewise(
             v_EI * 10 ** 12 / (self._E * self._I)
         )
-        
-        # THIS IS NOT CLEAR AND I DONT THINK IT IS CORRECT. FIX THIS. TO DO.
-        # Nv_EA represents the beam elongation, to make displacement need to add initial
-        # displacement. Can get the intial displacement by finding spring displacement
-        # at first support, and then minusing any axial deformation before it.
-        # Comparatively v_EI is already the beam displacement and has a constant in it
-        # (C2) that considers any intial displacement
-        if unknowns['x'][0]['stiffness'] == oo:
-            initial_spring_displacement_x = 0
-        else:
-            initial_spring_displacement_x = float(solutions_xx[0]  / unknowns['x'][0]['stiffness'])\
-        # in meters
-        self._axial_deflection= self.sympy_expr_to_piecewise(
-            (Nv_EA - NV_EA.subs(x, unknowns['x'][0]['position']))
-            * 10**3 / (self._E * self._A)
-            + initial_spring_displacement_x / 10 ** 3
-        )
 
     # SECTION - QUERY VALUE
     def get_reaction(self, x_coord, direction=None):
@@ -722,9 +700,10 @@ class Beam:
         else:
             return self._reactions[x_coord]
 
-    def _get_query_value(self,x_coord,sym_func,return_max=False,
-                         return_min=False,return_absmax=False):
-        """Find the value of a function at position x_coord.
+    def _get_query_value(self,x_coord, sym_func, return_max=False,
+                         return_min=False, return_absmax=False):
+        """Find the value of a function at position x_coord, or
+        determine function extremes.
 
         Parameters
         ----------
@@ -736,7 +715,7 @@ class Beam:
         return_max: bool
             return max value of function if true
         return_min: bool
-            return minx value of function if true
+            return min value of function if true
         return_absmax: bool
             return absolute max value of function if true
 
@@ -744,35 +723,54 @@ class Beam:
         --------
         int
             Max, min or absmax value of the sympy function depending
-            on which parameters are set.
+            on which parameters are set. If single x-coordinate set
+            then also returns int.
         list of ints
-            If x-coordinate(s) are specfied value of sym_func at
-            x-coordinate(s).
+            If x-coordinates are specfied value of sym_func at
+            x-coordinates. If max and min both True returns max and min,
+            and any x_coordinates that have been passed in (mainly useful
+            for Dash website result table generation speed.)
 
         Notes
         -----
[…]
