### 2021-02-24 31609db
Update indeterminatebeam.py
documentation update for indeterminatebeam. Mostly changing to PEP9 conventions.
