+++ indeterminatebeam/indeterminatebeam.py
@@ -1,4 +1,4 @@
-"""Main module that contains the main class for Beam 
+"""Main module that contains the main class for Beam
 and auxillary class for Support.
 
 Example
@@ -18,7 +18,7 @@ from math import radians
 
 # Third Party Imports
 import numpy as np
-from sympy import (integrate, lambdify, Piecewise, sympify, symbols, 
+from sympy import (integrate, lambdify, Piecewise, sympify, symbols,
                    linsolve, sin, cos, oo, SingularityFunction)
 from sympy.abc import x
 from plotly.subplots import make_subplots
@@ -38,21 +38,20 @@ from indeterminatebeam.loading import (
     DistributedLoad,
 )
 from indeterminatebeam.plotly_drawing_aid import (
-    draw_line, 
-    draw_arrowhead, 
-    draw_arrow, 
+    draw_line,
+    draw_arrowhead,
+    draw_arrow,
     draw_support_triangle,
-    draw_support_rectangle, 
-    draw_moment, 
-    draw_force, 
+    draw_support_rectangle,
+    draw_moment,
+    draw_force,
     draw_load_hoverlabel,
-    draw_reaction_hoverlabel, 
-    draw_support_hoverlabel, 
+    draw_reaction_hoverlabel,
+    draw_support_hoverlabel,
     draw_support_rollers,
-    draw_support_spring, 
+    draw_support_spring,
     draw_support
 )
-
 
 
 class Support:
@@ -111,7 +110,7 @@ class Support:
         """
         # validate coordinate
         assert_positive_number(coord, 'coordinate')
-        
+
         # validate support stiffness if assigned
         if kx:
             assert_positive_number(kx, 'kx')
@@ -198,7 +197,7 @@ class Beam:
         A sympy function representing the internal bending moments
         as a function of x.
     _deflection_equation: sympy piecewise function
-        A sympy function representing the tangential deflection as 
+        A sympy function representing the tangential deflection as
         a function of x.
 
     _reactions: dictionary of lists
@@ -273,37 +272,37 @@ class Beam:
         Parameters
         ----------
         *loads : iterable
-            An iterable containing load objects to be applied to the 
-            Beam object. Note that the load application point 
+            An iterable containing load objects to be applied to the
+            Beam object. Note that the load application point
             (or segment) must be within the Beam span.
 
         """
         # Iterate through each load passed into function
         for load in loads:
-            # Check if distributed load (will be true for V and H load 
+            # Check if distributed load (will be true for V and H load
             # types since they inherit from these main load types)
             if isinstance(load, (DistributedLoad, UDL, TrapezoidalLoad)):
-                
+
                 # check span exists within beam
                 left, right = load.span
                 if self._x0 > left or right > self._x1:
                     raise ValueError(
                         f"Coordinate {load.span} for {str(load)} is not a point on beam."
                     )
-                
+
                 # associate force variable with force value of load,
                 # in order to not add any loads that have no force later.
                 if isinstance(load, DistributedLoad):
                     # set force as arbritrary non-zero value so add any
                     # Distributed load to self._loads
-                    force = 1 
+                    force = 1
                 elif isinstance(load, UDL):
                     force = load.force
                 else:
                     force = max(load.force, key=abs)
 
             # check if point load/torque
-            elif isinstance(load,(PointTorque, PointLoad)):
+            elif isinstance(load, (PointTorque, PointLoad)):
 
                 # check point exists within beam
                 coordinate = load.position
@@ -312,17 +311,16 @@ class Beam:
                         f"Coordinate {coordinate} for {str(load)} is not a point on beam.")
 
                 force = load.force
-            
+
             # if not a distributed load or point load then isnt a load class.
             else:
[…]
