### 2021-02-24 8bcbca1
Analysis_reset and set_plotting_vectors methods
In indeterminate beam i added the analysis reset method. The purpose of the method is to make it clear what beam configuration results apply to. If a load is added to a previously analysed beam the analysis results should be reset to make sure the used doesnt think the properties they have only just set contributed to those previous results. As calculating data point vectors for plotting and for querys at seperate times was contributing to increases in speed, i instead made it so these vectors are only created when the beam is analysed and later 
[…]
