+++ indeterminatebeam/indeterminatebeam.py
@@ -258,13 +258,20 @@ class Beam:
         self._supports = []
         self._query = []
 
+        self._DATA_POINTS = 200
+
+        self.analysis_reset()
+
+    def analysis_reset(self):
+        """ Reset properties determined by analysis."""
+
         self._normal_forces = 0
         self._shear_forces = 0
         self._bending_moments = 0
         self._deflection_equation = 0
 
         self._reactions = {}
-        self._DATA_POINTS = 200
+        self._plotting_vectors = {}
 
     def add_loads(self, *loads):
         """Associate load objects with the beam object.
@@ -277,6 +284,8 @@ class Beam:
             (or segment) must be within the Beam span.
 
         """
+        self.analysis_reset()
+
         # Iterate through each load passed into function
         for load in loads:
             # Check if distributed load (will be true for V and H load
@@ -315,7 +324,7 @@ class Beam:
             # if not a distributed load or point load then isnt a load class.
             else:
                 raise ValueError(
-                    f"Load '{Load}' is not a load object.")
+                    f"Load '{load}' is not a load object.")
 
             # if force isnt 0, then add to self._loads, otherwise will
             # have no effect on the beam and as such shouldnt add.
@@ -335,6 +344,9 @@ class Beam:
             by default False.
 
         """
+
+        self.analysis_reset()
+
         # if remove all set to True, reintialize self._loads
         if remove_all:
             self._loads = []
@@ -350,6 +362,8 @@ class Beam:
         # an issue if they dont properly recreate the load they were
         # trying to remove and dont notice that they didnt actually
         # delete it.
+        
+
 
     def add_supports(self, *supports):
         """Associate support objects with the beam object.
@@ -362,6 +376,8 @@ class Beam:
             (or segment) must be within the Beam span.
 
         """
+
+        self.analysis_reset()
 
         # Check support valid then append to self._supports
         for support in supports:
@@ -404,6 +420,8 @@ class Beam:
             If true all supports associated with beam will be removed,
             by default False.
         """
+
+        self.analysis_reset()
 
         # if remove all set to True, reintialize self._supports
         if remove_all:
@@ -660,6 +678,44 @@ class Beam:
             v_EI * 10 ** 12 / (self._E * self._I)
         )
 
+        self.set_plotting_vectors()
+
+    def set_plotting_vectors(self):
+        """ Create vectors of data points for functions to 
+        allow for quicker plotting and determining of results."""
+        
+        x_vec = np.linspace(self._x0, self._x1, self._DATA_POINTS)
+
+        nf_func = lambdify(x, self._normal_forces, 'numpy')
+        sf_func = lambdify(x, self._shear_forces, 'numpy')
+        bm_func = lambdify(x, self._bending_moments, 'numpy')
+        d_func = lambdify(x, self._deflection_equation, 'numpy')
+
+        nf = np.array([float(nf_func(t)) for t in x_vec])
+        sf = np.array([float(sf_func(t)) for t in x_vec])
+        bm = np.array([float(bm_func(t)) for t in x_vec])
+        d = np.array([float(d_func(t)) for t in x_vec])
+
+        self._plotting_vectors = {
+            'x': x_vec,
+            'nf': {
+                'y_lam': nf_func,
+                'y_vec': nf,
+            },
+            'sf': {
+                'y_lam': sf_func,
+                'y_vec': sf,
+            },
+            'bm': {
+                'y_lam': bm_func,
+                'y_vec': bm,
+            },
+            'd': {
+                'y_lam': d_func,
+                'y_vec': d,
+            },
+        }
+
     # SECTION - QUERY VALUE
     def get_reaction(self, x_coord, direction=None):
         """Find the reactions of a support at position x.
[…]
