### 2021-02-24 1f88dcc
minor changes
plotly drawing aid update for plotting distributed type loads.
