+++ app.py
@@ -1013,167 +1013,170 @@ def analyse_beam(
     if input_json == prev_input:
         raise PreventUpdate
 
-    # try:
+    try:
 
-    t1 = time.perf_counter()
+        t1 = time.perf_counter()
 
-    if positive_y_direction == 'up':
-        d_ = 1
-    else:
-        d_ = -1
+        if positive_y_direction == 'up':
+            d_ = 1
+        else:
+            d_ = -1
 
-    for row in beams:
-        beam = Beam(*(float(a) for a in row.values()))
+        for row in beams:
+            beam = Beam(*(float(a) for a in row.values()))
 
-    if supports:
-        for row in supports:
-            if row['X'] in ['r', 'R']:
-                DOF_x = 1
-                kx = 0
-            elif row['X'] in ['f', 'F']:
-                DOF_x = 0
-                kx = 0
-            elif float(row['X']) > 0:
-                DOF_x = 0
-                kx = float(row['X'])
-            else:
-                raise ValueError(
-                    'input incorrect for x restraint of support')
+        beam._DATA_POINTS = data_points
 
-            if row['Y'] in ['r', 'R']:
-                DOF_y = 1
-                ky = 0
-            elif row['Y'] in ['f', 'F']:
-                DOF_y = 0
-                ky = 0
-            elif float(row['Y']) > 0:
-                DOF_y = 0
-                ky = float(row['Y'])
-            else:
-                raise ValueError(
-                    'input incorrect for y restraint of support')
+        if supports:
+            for row in supports:
+                if row['X'] in ['r', 'R']:
+                    DOF_x = 1
+                    kx = 0
+                elif row['X'] in ['f', 'F']:
+                    DOF_x = 0
+                    kx = 0
+                elif float(row['X']) > 0:
+                    DOF_x = 0
+                    kx = float(row['X'])
+                else:
+                    raise ValueError(
+                        'input incorrect for x restraint of support')
 
-            if row['M'] in ['r', 'R']:
-                DOF_m = 1
-            elif row['M'] in ['f', 'F']:
-                DOF_m = 0
-            else:
-                raise ValueError(
-                    'input incorrect for m restraint of support')
+                if row['Y'] in ['r', 'R']:
+                    DOF_y = 1
+                    ky = 0
+                elif row['Y'] in ['f', 'F']:
+                    DOF_y = 0
+                    ky = 0
+                elif float(row['Y']) > 0:
+                    DOF_y = 0
+                    ky = float(row['Y'])
+                else:
+                    raise ValueError(
+                        'input incorrect for y restraint of support')
 
-            beam.add_supports(
-                Support(
-                    float(row['Coordinate (m)']),
-                    (
-                        DOF_x,
-                        DOF_y,
-                        DOF_m
+                if row['M'] in ['r', 'R']:
+                    DOF_m = 1
+                elif row['M'] in ['f', 'F']:
+                    DOF_m = 0
+                else:
+                    raise ValueError(
+                        'input incorrect for m restraint of support')
+
+                beam.add_supports(
+                    Support(
+                        float(row['Coordinate (m)']),
+                        (
+                            DOF_x,
+                            DOF_y,
+                            DOF_m
+                        ),
+                        ky=ky,
+                        kx=kx,
                     ),
-                    ky=ky,
-                    kx=kx,
-                ),
-            )
-    # TO DO: add capitals
+                )
+        # TO DO: add capitals
 
-    if distributed_loads:
-        for row in distributed_loads:
-            if abs(float(row['Start Load (kN/m)'])) > 0 or \
-                    abs(float(row['End Load (kN/m)'])) > 0:
-                beam.add_loads(
[…]
