+++ docs/docstrings.rst
@@ -18,8 +18,6 @@ Beam
 .. autofunction:: indeterminatebeam.Beam.remove_loads
 .. autofunction:: indeterminatebeam.Beam.add_supports
 .. autofunction:: indeterminatebeam.Beam.remove_supports
-.. autofunction:: indeterminatebeam.Beam.get_support_details
-.. autofunction:: indeterminatebeam.Beam.check_determinancy
 .. autofunction:: indeterminatebeam.Beam.analyse
 .. autofunction:: indeterminatebeam.Beam.get_reaction
 .. autofunction:: indeterminatebeam.Beam.get_bending_moment
@@ -28,25 +26,21 @@ Beam
 .. autofunction:: indeterminatebeam.Beam.get_deflection
 .. autofunction:: indeterminatebeam.Beam.add_query_points
 .. autofunction:: indeterminatebeam.Beam.remove_query_points
-.. autofunction:: indeterminatebeam.Beam.plot
-.. autofunction:: indeterminatebeam.Beam.plot_beam_diagram
+.. autofunction:: indeterminatebeam.Beam.plot_beam_external
+.. autofunction:: indeterminatebeam.Beam.plot_beam_internal
+.. autofunction:: indeterminatebeam.Beam.plot_reaction_force
 .. autofunction:: indeterminatebeam.Beam.plot_normal_force
 .. autofunction:: indeterminatebeam.Beam.plot_shear_force
 .. autofunction:: indeterminatebeam.Beam.plot_bending_moment
 .. autofunction:: indeterminatebeam.Beam.plot_deflection
 
-PointTorque
-------------
-.. autoclass:: indeterminatebeam.PointTorque
+PointLoads
+-----------
+.. autoclass:: loading.PointTorque
+.. autoclass:: loading.PointLoad
 
-PointLoad
----------
-.. autoclass:: indeterminatebeam.PointLoad
-.. autoclass:: indeterminatebeam.PointLoadH
-.. autoclass:: indeterminatebeam.PointLoadV
-
-DistributedLoad
----------------
-.. autoclass:: indeterminatebeam.DistributedLoadV
-.. autoclass:: indeterminatebeam.DistributedLoadH
-.. autofunction:: indeterminatebeam.TrapezoidalLoad
+DistributedLoads
+-----------------
+.. autoclass:: loading.UDL
+.. autoclass:: loading.TrapezoidalLoad
+.. autoclass:: indeterminatebeam.DistributedLoad
