### 2021-02-25 19e055d
documentation update
sphinx update for read the docs. Slight docstring changes for code in docstrings. Updated examples and readme to use html instead of png for showing graphs in documentation. Fixed small bug for solving for moments in analysis as didnt consider y reactions in moment equilibrium.
