+++ docs/examples/example_1.py
@@ -17,7 +17,6 @@ beam.add_loads(load_1,load_2)
 
 beam.analyse()
 
-print(f"The beam has a determinancy of: {beam.check_determinancy()}")
 print(f"The beam has an absolute maximum shear force of: {beam.get_shear_force(return_absmax=True)} kN")
 print(f"The beam has an absolute maximum bending moment of: {beam.get_bending_moment(return_absmax=True)} kN.m")
 print(f"The beam has a vertical reaction at B of: {beam.get_reaction(3,'y')} kN")
