+++ docs/example.rst
@@ -31,7 +31,7 @@ The following values can be directly extracted using the `get_shear_force`, `get
 A plot of the axial force, shear force, and bending moments is shown below. A deflection graph is also presented however this depends on the beam properties E and I which werent included in this question.
 As a default the values E and I are taken as the values for a 150UB18.0 steel beam. 
 
-.. figure:: examples/example_1.png
+.. figure:: examples/example_1_internal.html
   :width: 700
   :alt: example_1
 
