+++ README.md
@@ -86,15 +86,15 @@ beam.analyse()                          #solves beam for unknowns
 ##### Plot results
 After the beam has been analysed we can plot the results.
 ```python
-beam.plot_beam_diagram()
+beam.plot_beam_external()
 beam.plot_beam_internal()                            
 ```
 
 The `plot` method is actually a wrapper that combines these four methods: `plot_normal_force`, `plot_shear_force`, `plot_bending_moment` and `plot_deflection` into a single A4-sized printer-friendly plot.
 
 The script above produces the following figures:
-![example_1 beam diagram plot](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example_diagram.png)
-![example_1 beam internal plot](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example_internal.png)
+![example_1 beam diagram plot](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example_external.html)
+![example_1 beam internal plot](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example_internal.html)
 
 
 ## Installing the package
