+++ indeterminatebeam/version.py
@@ -1,1 +1,1 @@
-__version__ = 'v1.2.3'
+__version__ = 'v2.0.0'
