### 2021-02-25 60c2c68
Update app.py
fixed instructions to have set width
