### 2021-02-26 339efef
minor changes
Updated hoverlabel for internal force plot. Added points slightly to either side of singularities for better plotting.
