+++ tool_comparison_table.tex
@@ -11,16 +11,16 @@
 \begin{tabular}{*{12}l}
 \multicolumn{1}{l}{ } & \mcrot{1}{l}{\rotangle}{Arbitrary distributed load functions} & \mcrot{1}{l}{\rotangle}{Full GUI (no code required)} & \mcrot{1}{l}{\rotangle}{Arbitrary number of loads} & \mcrot{1}{l}{\rotangle}{Free (as in free beer)} & \mcrot{1}{l}{\rotangle}{Open Source} & \mcrot{1}{l}{\rotangle}{Featured theory module} & \mcrot{1}{l}{\rotangle}{Detailed solution procedure} & \mcrot{1}{l}{\rotangle}{Programmable interface} & \mcrot{1}{l}{\rotangle}{Spring Supports} & \mcrot{1}{l}{\rotangle}{Any DOF combination for Supports} & \mcrot{1}{l}{\rotangle}{Any number of supports} \\ %\hline
 \cmidrule{2-12}
-\href{https://github.com/JesseBonanno/IndeterminateBeam/}{IndeterminateBeam}  	& \cmark & \xmark & \cmark & \cmark & \cmark & \cmark & \xmark & \cmark & \cmark & \cmark & \cmark  \\%\hline
-\href{https://alfredocarella.github.io/simplebendingpractice/}{BeamBending} 		& \cmark & \xmark & \cmark & \cmark & \cmark & \cmark & \xmark & \cmark & \xmark & \xmark & \xmark \\ %\hline
-\href{http://rascheta.net/beamuk/}{Beam Calculator Online}				& \xmark & \cmark & \cmark & \cmark & \xmark & \xmark & \xmark & \xmark & \xmark & \xmark & \xmark \\ %\hline
-\href{http://beamguru.com/}{Beam Guru} 							& \xmark & \cmark & \cmark & \xmark & \xmark & \xmark & \cmark & \xmark  & \xmark & \xmark & \cmark  \\ %\hline
-\href{https://mechanicalc.com/calculators/beam-analysis/}{MechaniCalc}			& \xmark & \cmark & \cmark & \xmark & \xmark & \cmark & \xmark & \xmark  & \xmark & \cmark & \cmark \\ %\hline
-\href{https://skyciv.com/structural-software/beam-analysis-software/}{SkyCiv Beam} & \xmark & \cmark & \cmark & \xmark & \xmark & \xmark & \cmark & \xmark  & \xmark & \xmark & \xmark  \\ %\hline
-\href{https://www.steelbeamcalculator.com/}{Steel Beam Calculator} 			& \xmark & \cmark & \cmark & \xmark & \xmark & \xmark & \xmark & \xmark  & \xmark & \xmark & \xmark  \\ %\hline
-\href{https://www.amesweb.info/StructuralBeamDeflection/BeamDeflectionCalculators.aspx}{Structural Beam Calculator}& \xmark & \cmark & \xmark & \cmark & \xmark & \xmark & \xmark & \xmark  & \xmark & \xmark & \xmark  \\ %\hline
-\href{https://webstructural.com/shear-and-moment-diagram.html}{WebStructural} 	& \xmark & \cmark & \cmark & \xmark & \xmark & \xmark & \xmark & \xmark  & \xmark & \xmark & \cmark  \\ %\hline
-\href{https://clearcalcs.com/freetools/beam-analysis/au}{ClearCalcs} 			& \xmark & \cmark & \cmark & \xmark & \xmark & \xmark & \xmark & \xmark  & \cmark & \xmark & \cmark  \\ %\hline
[…]
