### 2021-02-26 2d1d9ff
_get_query_value update
Addresses issues #17 and #18
