+++ README.md
@@ -1,7 +1,7 @@
 # Indeterminate Beam
 
 
-[![Version](https://img.shields.io/badge/version-v2.0.0-blue.svg)](https://github.com/JesseBonanno/IndeterminateBeam/releases/tag/v2.0.0)
+[![Version](https://img.shields.io/badge/version-v2.0.0-blue.svg)](https://github.com/JesseBonanno/IndeterminateBeam/releases/tag/v2.0.1)
 [![License](https://img.shields.io/badge/license-MIT-lightgreen.svg)](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/LICENSE.txt)
 [![Documentation Status](https://readthedocs.org/projects/indeterminatebeam/badge/?version=main)](https://indeterminatebeam.readthedocs.io/en/main/?badge=main)
 [![Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/simple_demo.ipynb)
