### 2021-02-26 0ed40d2
Update indeterminatebeam.py
