### 2021-02-26 0045dc8
jupyter update
Updated Jupyter notebook to better reflect the latest code changes.
