### 2021-02-27 a1eefb0
favicon
added asset folder with favicon
