+++ app.py
@@ -900,7 +900,7 @@ content_first_row = html.Div(
 
 content = html.Div(
     [
-        html.H2('Beam Calculator', style=TEXT_STYLE),
+        html.H2('IndeterminateBeam Calculator', style=TEXT_STYLE),
         html.Hr(),
         calc_status,
         content_first_row,
