+++ assets/about.txt
@@ -1,0 +1,6 @@
+This favicon was generated using the following font:
+
+- Font Title: Bigshot One
+- Font Author: Copyright (c) 2011 by Gesine Todt. All rights reserved.
+- Font Source: http://fonts.gstatic.com/s/bigshotone/v11/u-470qukhRkkO6BD_7cM_gxuUQJBXv_-.ttf
+- Font License: SIL Open Font License, 1.1 (http://scripts.sil.org/OFL)
