+++ docs/example.rst
@@ -5,10 +5,13 @@ Indeterminate Beam with Point Load and Distributed Load
 
 This example demonstrates some functionality of the ``indeterminatebeam`` package. It has been taken from example 12.21 of the Hibbeler textbook :cite:`HibbelerRussell2013MoM`.
 
-Try it in online: |colab|
+Try it in online: |colab| |website|
 
 .. |colab| image:: https://colab.research.google.com/assets/colab-badge.svg
    :target: https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/example_1.ipynb
+
+.. |website| image:: https://img.shields.io/website?down_color=lightgrey&down_message=offline&up_color=green&up_message=up&url=https%3A%2F%2Findeterminate-beam.herokuapp.com%2F
+   :target: https://indeterminate-beam.herokuapp.com/
 
 Specifications
 --------------
