+++ README.md
@@ -1,10 +1,11 @@
 # Indeterminate Beam
 
 
-[![Version](https://img.shields.io/badge/version-v2.0.0-blue.svg)](https://github.com/JesseBonanno/IndeterminateBeam/releases/tag/v2.0.1)
+[![Version](https://img.shields.io/badge/version-v2.0.1-blue.svg)](https://github.com/JesseBonanno/IndeterminateBeam/releases/tag/v2.0.1)
 [![License](https://img.shields.io/badge/license-MIT-lightgreen.svg)](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/LICENSE.txt)
 [![Documentation Status](https://readthedocs.org/projects/indeterminatebeam/badge/?version=main)](https://indeterminatebeam.readthedocs.io/en/main/?badge=main)
 [![Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/simple_demo.ipynb)
+[![Website](https://img.shields.io/website?down_color=lightgrey&down_message=offline&up_color=green&up_message=up&url=https%3A%2F%2Findeterminate-beam.herokuapp.com%2F)](https://indeterminate-beam.herokuapp.com/)
 
 IndeterminateBeam is a Python Package aiming to serve as a foundation for civil and structural engineering projects in Python. The module can also serve as a stand-alone program and is useful for determining:
 
