### 2021-02-28 eba8531
update app.py for report generation
Reports generate now fit better on two pages. Can fit up to 6 query points before rolls into 3 pages.
