### 2021-02-28 073b9ae
Travis CI
Added Travis CI yml file, and added badge to readme for travis CI. Fixed errors in unittest file.
