+++ requirements.txt
@@ -13,3 +13,4 @@ dash-renderer==1.8.3
 dash-table==4.11.1
 gunicorn
 dash-extensions>=0.0.44
+pytest
