+++ .travis.yml
@@ -18,6 +18,7 @@ branches:
 
 install:
   - pip install -r requirements.txt
+  - pip install .
 
 # set up the display port to generate figures
 # ref: https://docs.travis-ci.com/user/gui-and-headless-browsers/#using-xvfb-directly
