+++ app.py
@@ -36,6 +36,12 @@ SIDEBAR_STYLE = {
     'padding': '40px 40px',
     'background-color': '#f8f9fa'
 }
+from indeterminatebeam import PointLoadV, PointTorque, DistributedLoadV
+load_1 = PointLoadV(1,2) # Create 1kN point load at x = 2m
+load_2 = DistributedLoadV(2,(1,4)) # Create a 2kN UDL from x = 1m to x = 4m
+load_3 = PointTorque(2, 3.5) # Defines a 2kN.m point torque at x = 3.5m
+beam.add_loads(load_1,load_2,load_3) # Assign the load objects to the beam object
+
 
 # the style arguments for the main content page.
 CONTENT_STYLE = {
@@ -70,8 +76,7 @@ For more, you can view the following:
    The sign conventions used
 * [![Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/simple_demo.ipynb)
    The Python based Jupyter Notebook examples
-* [![Article](https://img.shields.io/badge/Article-Submitted-orange.svg)](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/paper.md)
-   The JOSE article concerning this package
+* [![Article](https://img.shields.io/badge/Article-Complete-green.svg)](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/Article.pdf)
 
 Note: As the Python package calculations are purely analytical calculation \
 times can be relatively slow.
