+++ tool_comparison_table.tex
@@ -9,7 +9,7 @@
 \begin{document}
 \begin{table}[]
 \begin{tabular}{*{12}l}
-\multicolumn{1}{l}{ } & \mcrot{1}{l}{\rotangle}{Arbitrary distributed load functions} & \mcrot{1}{l}{\rotangle}{Full GUI (no code required)} & \mcrot{1}{l}{\rotangle}{Arbitrary number of loads} & \mcrot{1}{l}{\rotangle}{Free (as in free beer)} & \mcrot{1}{l}{\rotangle}{Open Source} & \mcrot{1}{l}{\rotangle}{Featured theory module} & \mcrot{1}{l}{\rotangle}{Detailed solution procedure} & \mcrot{1}{l}{\rotangle}{Programmable interface} & \mcrot{1}{l}{\rotangle}{Spring Supports} & \mcrot{1}{l}{\rotangle}{Any DOF combination for Supports} & \mcrot{1}{l}{\rotangle}{Any number of supports} \\ %\hline
+\multicolumn{1}{l}{ } & \mcrot{1}{l}{\rotangle}{Arbitrary distributed load functions} & \mcrot{1}{l}{\rotangle}{Full GUI (no code required)} & \mcrot{1}{l}{\rotangle}{Arbitrary number of loads} & \mcrot{1}{l}{\rotangle}{Free} & \mcrot{1}{l}{\rotangle}{Open Source} & \mcrot{1}{l}{\rotangle}{Featured theory module} & \mcrot{1}{l}{\rotangle}{Detailed solution procedure} & \mcrot{1}{l}{\rotangle}{Programmable interface} & \mcrot{1}{l}{\rotangle}{Spring Supports} & \mcrot{1}{l}{\rotangle}{Any DOF combination for Supports} & \mcrot{1}{l}{\rotangle}{Any number of supports} \\ %\hline
 \cmidrule{2-12}
 {IndeterminateBeam}  	& \cmark & \cmark & \cmark & \cmark & \cmark & \cmark & \xmark & \cmark & \cmark & \cmark & \cmark  \\%\hline
 {BeamBending} 		& \cmark & \xmark & \cmark & \cmark & \cmark & \cmark & \xmark & \cmark & \xmark & \xmark & \xmark \\ %\hline
