+++ article.md
@@ -1,186 +1,0 @@
-# IndeterminateBeam: A Python package for solving 1D indeterminate beams
-Jesse Bonanno
-04 March 2021
-
-## Summary
-
-`IndeterminateBeam` is a Python package aiming to serve as a foundation for civil and structural engineering projects in Python. The package can also serve as a standalone program and is useful for determining: 
-
-  - reaction forces for indeterminate beams
-  - internal forces for indeterminate beams (shear, bending, axial)
-  - deflection of the beam due to resulting forces
-  - axial force, shear force, and bending moment diagrams
-
-The package is based mainly on engineering concepts of statics as described in (Hibbeler, 2013), and python packages Sympy (Meurer et al., 2017) and Plotly (Plotly Technologies Inc, 2015). The [package documentation](https://indeterminatebeam.readthedocs.io/en/main/) provides a brief overview of the theory behind the solutions used to calculate the forces on the indeterminate beam.
-
-The `IndeterminateBeam` [package repository](https://github.com/JesseBonanno/IndeterminateBeam) can be found on Github and is ready for installation using `pip`. A text based example of the package can be found on this [Jupyter notebook](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.ipynb) and a web-based graphical user interface (GUI) is available at https://indeterminate-beam.herokuapp.com/.
-
-
-## Project Purpose
-
-The purpose of this project was two-fold:
-1.	Create a free website that has more features than paid and free beam calculators that exist on the web.
-2.	Provide a foundation for civil and structural engineers who want to create higher order engineering python programs.
-
-Several (mostly paid) beam calculator websites currently exist online, providing the same service as this package, with web traffic in the hundreds of thousands per month (Similiarweb, 2021). Despite this, no online service exists (to the authors knowledge) that has all the features of IndeterminateBeam and is also free.
-
- Similiarly, there are no well-documented indeterminate beam solving Python packages (to the authors knowledge) despite the importance of such a calculation in engineering. Several python finite element analysis (FEA) packages do exist, however they are vastly overcomplicated for someone wanting to only solve for forces on a one-dimensional beam.
-
-This python package was heavily inspired by [beambending](https://github.com/alfredocarella/simplebendingpractice) (Carella, 2019), a module created for educational purposes by Alfredo Carella of the Oslo Metropolitan University. The beambending module, although well documented, can only solve for simply supported beams consisting of a pin and roller support. The [package documentation](https://simplebendingpractice.readthedocs.io/en/latest/?badge=latest)  for this project includes a more rigorous overview of the theory behind the basics for solving determinate structures.
-
-## Project Comparison
-A feature comparison in Table 1 below has been taken from (Carella, 2019) and modified to include more packages and features.
-
-![Table 1](https://github.com/JesseBonanno/IndeterminateBeam/raw/main/tool_comparison_table.png)
- 
-There are five main differences between the ```IndeterminateBeam``` package and the rest of the reviewed tools:
-
-* Arbitrary distributed load functions are accepted (as long as sympy can parse them), i.e. they are not restricted to constants or linear functions.
-* The package can be called from regular Python code, which makes it easy to implement higher order engineering solutions.
-* It is not only free but also completely open-source.
-* Spring supports can be modelled
[…]
