### 2021-03-19 06006d2
Persistent data and uploads
Address #20. Added functionality to store data from previous session, functionality to reset all data and functionality to load data from a previous report. Involved combining all these features into a single function since DASH only allows an output to occur once. Hence anything involving updating a table was combined.
