+++ app.py
@@ -1,5 +1,6 @@
 import sys
 import os
+import base64
 import json
 import dash
 import dash_bootstrap_components as dbc
@@ -147,8 +148,6 @@ beam_table = dash_table.DataTable(
     data=[beam_table_init],
     editable=True,
     row_deletable=False,
-    persistence=True,
-    persistence_type='session'
 )
 
 beam_instructions = dcc.Markdown('''
@@ -558,7 +557,7 @@ distributed_load_content = dbc.Card(
 )
 
 # Properties for query tab
-query_table_data = {
+query_table_init = {
     'Query coordinate (m)': 0,
 }
 
@@ -573,10 +572,10 @@ query_table = dash_table.DataTable(
         'format': Format(
                 symbol=Symbol.yes,
                 symbol_suffix=' m')
-    } for i in query_table_data.keys()],
+    } for i in query_table_init.keys()],
     data=[],
     editable=True,
-    row_deletable=True
+    row_deletable=True,
 )
 
 query_instructions = dcc.Markdown('''
@@ -717,10 +716,10 @@ option_positive_direction_y = dbc.FormGroup(
 
 option_result_table = dbc.FormGroup(
     [
-        dbc.Label("Result Table", html_for="option-result-table", width=3),
+        dbc.Label("Result Table", html_for="option_result_table", width=3),
         dbc.Col(
             dbc.RadioItems(
-                id="option-result-table",
+                id="option_result_table",
                 options=[
                     {'label': 'Hide', 'value': 'hide'},
                     {'label': 'Show', 'value': 'show'},
@@ -749,7 +748,7 @@ option_data_point = dbc.FormGroup(
                     250: {'label': '250'},
                     500: {'label': '500'}
                 },
-                included=True
+                included=True,
             ),
             width=8,
         ),
@@ -762,7 +761,33 @@ option_combined = dbc.Form([
     option_support_input,
     option_positive_direction_y,
     option_data_point,
-
+    html.Br(),
+    dbc.Row(
+        [
+            dbc.Label("Work from Report", width=3),
+            html.Div([
+                dcc.Upload(
+                    id='upload-data',
+                    children=html.Div([
+                        'Drag and Drop or ',
+                        html.A('Select Files'),
+                    ]),
+                    style={
+                        'width': '100%',
+                        'height': '60px',
+                        'lineHeight': '60px',
+                        'borderWidth': '1px',
+                        'borderStyle': 'dashed',
+                        'borderRadius': '5px',
+                        'textAlign': 'center',
+                        'margin': '5px'
+                    },
+                    # Allow multiple files to be uploaded
+                    multiple=False
+                ),
+            ])
+        ]
+    )
 ])
 
 option_content = dbc.Card(
@@ -874,10 +899,25 @@ content_first_row = html.Div(
                                 width=4
                             ),
                             dbc.Col(
-                                dbc.Spinner(submit_button),
+                                [
+                                    dbc.Button(
+                                        "Clear Inputs",
+                                        id="clear-inputs-button",
+                                        className="mb-3",
+                                        color="info",
+                                        n_clicks=0,
+                                        block=True,
+                                    ),
+                                ],
                                 width=4
-                            )
-                        ]
+                            ),
+                        ],
+                    ),
+                    dbc.Row(
+                        dbc.Col(
+                            dbc.Spinner(submit_button),
[…]
