### 2021-04-01 #107 @moorepants
@alfredocarella Would you be interested in reviewing this submission? It is very similar to your prior JOSE article.
