### 2021-04-01 #107 @moorepants
@amcc1996 I see that you've developed the symbeam package. I believe you'd be an appropriate reviewer for this Journal of Open Source Education submission on a similar topic. Would you like to be a reviewer?
