### 2021-04-01 #107 @JesseBonanno
If @alfredocarella is able to review i think he would be very well suited for the task. I have privately been in contact with him and he has already had a look at the code and given feedback so he is halfway into the role already.

Other potential candiates i can find are Satish Annigeri, Kyle Niemeyer, Cameron Devine (in order of suitability). Chosen for a background in mechanical/civil/structural engineering with knowledge of python.

I would also like to note that since i submitted this paper i made some changes to calculation method based on recommendations from Alfredo Carella. This has led me to a more pythonic code at version v2.0.1 which i am more satsified with. Would i be able to update this review from version v1.2.2 to version v2.0.1 or is that too difficult?
