### 2021-04-02 #107 @moorepants
> This has led me to a more pythonic code at version v2.0.1 which i am more satsified with. Would i be able to update this review from version v1.2.2 to version v2.0.1 or is that too difficult?

Yes please update.
