### 2021-04-02 #107 @JesseBonanno
I think i am going to have to update the paper to reflect some minor changes to referenced images etc, that i didnt account for during last few commits. I will update the paper now, shouldn't take too long.
