+++ paper.md
@@ -17,7 +17,7 @@ authors:
 affiliations:
 - name: No affiliation
   index: 1
-date: 15 December 2020
+date: 02 April 2021
 bibliography: paper.bib
 nocite: | 
   @beamguru, @structuralbeam, @skyciv, @mechanicalc, @webstructural, @beamcalculatoronline, @steelbeamcalculator
@@ -25,12 +25,12 @@ nocite: |
 
 # Summary
 
-IndeterminateBeam is a Python package aiming to serve as a foundation for civil and structural engineering projects in Python. The module can also serve as a stand-alone program and is useful for determining:
+`IndeterminateBeam` is a Python package aiming to serve as a foundation for civil and structural engineering projects in Python. The package can also serve as a standalone program and is useful for determining: 
 
   - reaction forces for indeterminate beams
   - internal forces for indeterminate beams (shear, bending, axial)
-  - deflections of beams due to resulting forces
-  - axial force, shear force, bending moment and deflection diagrams.
+  - deflection of beams due to resulting forces
+  - axial force, shear force, bending moment and deflection diagrams
 
 The module is based primarily on engineering concepts of statics as described in [@HibbelerRussell2013MoM], and Python packages Sympy [@sympy2017] and Matplotlib [@Hunter2007]. The [package documentation](https://indeterminatebeam.readthedocs.io/en/main/) provides a brief overview of the theory behind the solutions used to calculate the forces on the indeterminate beam.
 
@@ -41,9 +41,7 @@ The package can be used by:
    - students who want to observe the effect minor geometry changes can have on forces
    - students, teachers or engineers who want to create higher order engineering projects using this project as a starting point
 
-
-The `indeterminatebeam` package is ready for installation using `pip` or can be tested online using the provided [Jupyter notebook](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.ipynb).
-
+The `IndeterminateBeam` [package repository](https://github.com/JesseBonanno/IndeterminateBeam) can be found on Github and is ready for installation using `pip`. A text-based example of the package can be found on this [Jupyter Notebook](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.ipynb) and a web-based graphical user interface (GUI) is available at https://indeterminate-beam.herokuapp.com/.
 
 # Statement of Need
 Statics is fundamental to many fields of engineering such as civil, structural and mechanical engineering. This package aims to help student understanding in two ways:
@@ -65,41 +63,38 @@ There are five main differences between the ```indeterminatebeam``` package and 
 * Any degree of freedom combination can be constructed for supports 
 
 
+## Functionality and Usage
 
-# Functionality and Usage
-
-A typical use case of the `indeterminatebeam` package involves the following steps:
+A typical use case of the `IndeterminateBeam` package involves the following steps:
 
 1. Create a `Beam` object
 2. Create `Support` objects and assign to `Beam`
-3. Create `load` objects and assign to `Beam`
+3. Create `Load` objects and assign to `Beam`
 4. Solve for forces on `Beam` object
 5. Plot results
 
-Units and load direction conventions are described in the [package documentation](https://indeterminatebeam.readthedocs.io/en/main/).
+You can follow along with the example below in this web-based [Jupyter Notebook](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.ipynb). Units and load direction conventions are described in the [package documentation](https://indeterminatebeam.readthedocs.io/en/main/theory.html).
 
[…]
