### 2021-04-02 #107 @alfredocarella
@moorepants I appreciate being considered as a reviewer. However, as @JesseBonanno  correctly points out, we have exchanged a few emails and I have privately suggested improvements. I am afraid this might appear as a potential conflict of interest, and believe it would be healthy to avoid it if possible.
In case you don't find a second reviewer within reasonable time, I am willing to take the role as a last resource. I just believe I should not be the first choice.
