### 2021-04-02 #107 @moorepants
@satish-annigeri would you be interested in reviewing this Journal of Open Source Education submission regarding beam bending structural mechanics?
