### 2021-04-02 #107 @JesseBonanno
Sounds fair @alfredocarella, i hadn't considered that.

Also, I have updated the paper to be completely in-sync with any changes from v2.0.1.
I will generate the paper below now.
