### 2021-04-02 #107 @moorepants
@JesseBonanno From a quick pass in your paper, I see that you use SymPy in your tool but don't include the builtin beam bending functionality in your comparison table. I think it could be worth adding that. Here are the docs: https://docs.sympy.org/latest/modules/physics/continuum_mechanics/index.html
