### 2021-04-02 #107 @amcc1996
While we wait to start the more in-depth review process, I skimmed through the contents of the repository to check if something popped out to me.

I completely agree with @moorepants. Without wanting to raise any conflicts of interest here, in my opinion, the article will benefit from a more up-to-date version of the beam bending software out there. As this work relies on a Python implementation, maybe one should consider including the recent alternative [SymBeam](https://github.com/amcc1996/symbeam) in the comparison table, for the sake of completeness. To my knowledge, I do not know of any other Python alternatives, which ultimately motivates me to make this suggestion. Do you think this is reasonable @moorepants , or it grasps the limits of conflict of interests?

Regarding the [`README.md`](README.md) file, I think there is a minor typo in the *Plot results* section. In the sentence *The `plot` method is actually a wrapper that combines these four methods*, it seems that you should be referring to the `plot_beam_internal` method instead. Can you look at it @JesseBonanno?

Last, maybe I am being too peaky, but for those not so acquainted with `plotly`, it could advantageous to include the plotting lines `fig.show()` in the example described in the repository. Otherwise, one may end running `python3 intermediatebeam.py` a number of times an getting absolutely no output. What to do you think @JesseBonanno?
