### 2021-04-02 #107 @moorepants
I think it is reasonable to mention closely related packages, regardless of the conflict of interest, but it's fine if they are left out too.
