### 2021-04-03 #107 @JesseBonanno
@moorepants I agree that i should add SymPy in.

@amcc1996 I don't have any issues adding in SymBeam. I think it would be dishonest to not include anything with similar capabilities.

And you are correct that is a typo, and i agree i should add in fig.show().

I will make these changes within the next few days.
