### 2021-04-07 ed85777
Paper Update
Updated images to be higher definition. Added in SymPy and SymBeam into comparison table, and added into bibliography.
