+++ paper.bib
@@ -66,6 +66,14 @@ keywords = {Structural analysis (Engineering) ; Strength of materials},
  doi = {10.7717/peerj-cs.103}
 }
 
+@misc{symbeam,
+  title = {{SymBeam}},
+  year = {2021},
+  author = {António Carneiro}
+  howpublished = {\url{https://pypi.org/project/symbeam/}},
+  publisher = {{Accessed on 2020-04-07}}
+}
+
 @misc{beamguru,
   title = {{Beam Guru}},
   year = {2020},
