+++ tool_comparison_table.tex
@@ -13,6 +13,8 @@
 \cmidrule{2-12}
 {IndeterminateBeam}  	& \cmark & \cmark & \cmark & \cmark & \cmark & \cmark & \xmark & \cmark & \cmark & \cmark & \cmark  \\%\hline
 {BeamBending} 		& \cmark & \xmark & \cmark & \cmark & \cmark & \cmark & \xmark & \cmark & \xmark & \xmark & \xmark \\ %\hline
+{SymBeam}  	& \cmark & \cmark & \cmark & \cmark & \cmark & \cmark & \cmark & \cmark & \xmark & \xmark & \xmark  \\%\hline
+{SymPy}  	& \cmark & \xmark & \cmark & \cmark & \cmark & \xmark & \xmark & \cmark & \xmark & \cmark & \cmark  \\%\hline
 {Beam Calculator Online}				& \xmark & \cmark & \cmark & \cmark & \xmark & \xmark & \xmark & \xmark & \xmark & \xmark & \xmark \\ %\hline
 {Beam Guru} 							& \xmark & \cmark & \cmark & \xmark & \xmark & \xmark & \cmark & \xmark  & \xmark & \xmark & \cmark  \\ %\hline
 {MechaniCalc}			& \xmark & \cmark & \cmark & \xmark & \xmark & \cmark & \xmark & \xmark  & \xmark & \cmark & \cmark \\ %\hline
