+++ paper.bib
@@ -69,7 +69,7 @@ keywords = {Structural analysis (Engineering) ; Strength of materials},
 @misc{symbeam,
   title = {{SymBeam}},
   year = {2021},
-  author = {António Carneiro}
+  author = {António Carneiro},
   howpublished = {\url{https://pypi.org/project/symbeam/}},
   publisher = {{Accessed on 2020-04-07}}
 }
