+++ paper.bib
@@ -71,7 +71,7 @@ keywords = {Structural analysis (Engineering) ; Strength of materials},
   year = {2021},
   author = {António Carneiro},
   howpublished = {\url{https://pypi.org/project/symbeam/}},
-  publisher = {{Accessed on 2020-04-07}}
+  publisher = {{Accessed on 2021-04-07}}
 }
 
 @misc{beamguru,
