### 2021-04-07 #107 @moorepants
@Psycho-Pirate would you be interested in reviewing this Journal of Open Source Education submission regarding beam bending structural mechanics?
