### 2021-04-07 #111 @whedon
[REVIEW]: IndeterminateBeam: A Python package for solving 1D indeterminate beams
**Submitting author:** @jessebonanno (<a href="http://orcid.org/0000-0002-4996-6813">Jesse Bonanno</a>)
**Repository:** <a href="https://github.com/JesseBonanno/IndeterminateBeam/" target ="_blank">https://github.com/JesseBonanno/IndeterminateBeam/</a>
**Version:** v2.1.6
**Reviewer:** @amcc1996 , @nicoguaro
