### 2021-04-07 #107 @Psycho-Pirate
> @Psycho-Pirate would you be interested in reviewing this Journal of Open Source Education submission regarding beam bending structural mechanics?

Hey Jason, I would definitely be interested in reviewing it. However, I am currently occupied with preparing my GSoC proposal for Sympy. I will be available after few days. Is there any deadline for this?
@moorepants
