### 2021-04-07 #111 @moorepants
@JesseBonanno, @amcc1996 and @nicoguaro the review will occur here. Anywhere you see JOSS above, take it as JOSE (as we share the review system with JOSS). But the same guidelines apply. @amcc1996 and @nicoguaro make sure to review the documentation for being a reviewer here: https://openjournals.readthedocs.io/en/jose/. Let me know if you have any questions.

How does three weeks for the first pass for review sound? If not good, please propose a duration that can work for you.
