### 2021-04-07 97f33d7
CodeFactor and CodeCoverage
Added codefactor and codecoverage badges. Made changes to travis.yml for codecoverage. Made small changes to code based on current issues raised by codefactor.
