+++ tests/test_indeterminatebeam.py
@@ -1,4 +1,5 @@
-import sys, os
+import sys
+import os
 sys.path.insert(0, os.path.abspath('../'))
 
 from indeterminatebeam.indeterminatebeam import (
