### 2021-04-07 #111 @nicoguaro
@JesseBonanno, I followed the installation as recommended in the README file and tried the example in the documentation but did not manage to get the plots. When I ran it in Visual Studio Code I got the following error message

> ValueError: Mime type rendering requires nbformat>=4.2.0 but it is not installed
