### 2021-04-07 f43a69b
Update .travis.yml
trying to configure codecoverage
