### 2021-04-08 #111 @JesseBonanno
> @JesseBonanno, I followed the installation as recommended in the README file and tried the example in the documentation but did not manage to get the plots. When I ran it in Visual Studio Code I got the following error message
>
> > ValueError: Mime type rendering requires nbformat>=4.2.0 but it is not installed

@nicoguaro I took a look at the google Golab and Binder online examples and they didn't have this issue. To try resolve the issue I have added in jupyter>=1.0.0 into the requirements.txt which I believe should work, although I am having difficulty recreating this issue at the moment. Please try the installation again and let me know if it works for you.
