### 2021-04-08 #111 @nicoguaro
> @nicoguaro I took a look at the google Golab and Binder online examples and they didn't have this issue. To try resolve the issue I have added in jupyter>=1.0.0 into the requirements.txt which I believe should work, although I am having difficulty recreating this issue at the moment. Please try the installation again and let me know if it works for you.

@JesseBonanno, I don't think that work. I am using the suggested method of installation

    pip install indeterminate beams

You would need to create a new version there.

I could try installing from the cloned repository, but I first try to use the instructions verbatim since that is what the users will try.
