### 2021-04-08 #111 @JesseBonanno
@nicoguaro oh right, yes i would need to update the version.

For now could you try pip install jupyter>=1.0.0 alongside the pip install indeterminatebeam so i can verify if this fixes the issue you had since i am having trouble recreating this issue.

If it works i will push that version to 2.0.2 and have whedon update the version.
