### 2021-04-08 #111 @JesseBonanno
I have published v2.0.2 to ammend the vscode jupyter notebook issue as described in the previous few comments.

Can we please change the version to v2.0.2 to capture this?
