### 2021-04-08 #111 @nicoguaro
@JesseBonanno I tried in Visual Studio code and still have some problems, that are solved after installing nbformat. Your change make the plots works, though. I tried them in plain Python and Jupyter notebook.

Regarding the software, it is giving wrong results. Even the example in the README is wrong since it does not satisfy equilibrium in the vertical direction.
