### 2021-04-08 #111 @JesseBonanno
@nicoguaro sorry for not noticing this before the submission, this is obviously a big issue and thank you for noticing it.

I compared the solution to the solution from 1.2.3, and I believe the previous solution was correct.

Previous: https://github.com/JesseBonanno/IndeterminateBeam/blob/v1.2.3/docs/examples/readme_example_internal.png
Current: https://github.com/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example_internal_HD.png

There seems to be an issue with the interpretation of starting and ending points for the UDL's in the update version 2, which hopefully won't take too long to amend (if the UDL is considered to run the full length then i think the solution would have been correct). I will also try to make my unit tests more robust as this wasn't captured in my unit test.

I will fix this up on the weekend.
