### 2021-04-11 9c777a2
Bug Fix DistributedLoad
Bug fix for Distributed Load. Update to Jupyter-Notebook demos to be more comprehensive.
