+++ docs/examples/readme_example.py
@@ -1,5 +1,7 @@
 # arbritrary example defined in README.md
-from indeterminatebeam import Beam, Support, PointLoadV, PointTorque, DistributedLoadV
+import sys, os
+sys.path.insert(0, os.path.abspath('../../'))
+from indeterminatebeam import Beam, Support, PointLoadV, PointTorque, UDLV, DistributedLoadV
 beam = Beam(7)                          # Initialize a Beam object of length 9m with E and I as defaults
 beam_2 = Beam(9,E=2000, I =100000)      # Initializa a Beam specifying some beam parameters
 
@@ -9,15 +11,21 @@ c = Support(7,(1,1,1))                  # Defines a fixed support at location x 
 beam.add_supports(a,b,c)    
 
 load_1 = PointLoadV(1,2)                # Defines a point load of 1kn acting up, at location x = 2m
-load_2 = DistributedLoadV(2,(1,4))      # Defines a 2kN UDL from location x = 1m to x = 4m 
-load_3 = PointTorque(2, 3.5)              # Defines a 2kN.m point torque at location x = 3.5m
-beam.add_loads(load_1,load_2,load_3)           # Assign the support objects to a beam object created earlier
+load_2 = UDLV(2,(1,4))                  # Defines a 2kN UDL from location x = 1m to x = 4m 
+load_3 = PointTorque(2, 3.5)            # Defines a 2kN.m point torque at location x = 3.5m
+beam.add_loads(load_1,load_2,load_3)    # Assign the support objects to a beam object created earlier
 
-beam.analyse()
+beam.analyse() 
 
-fig_1 = beam.plot_beam_diagram()
+fig_1 = beam.plot_beam_external()
 fig_1.show()
 
 fig_2 = beam.plot_beam_internal()
 fig_2.show()
-#fig.write_image("./readme_example.png")
+
+# save the results (optional)
+# Can save figure using ``fig.write_image("./results.pdf")`` (can change extension to be
+# png, jpg, svg or other formats as reired). Requires pip install -U kaleido
+
+# fig_1.write_image("./readme_example_diagram.png")
+# fig_2.write_image("./readme_example_internal.png")
