+++ README.md
@@ -1,7 +1,7 @@
 # Indeterminate Beam
 
 
-[![Version](https://img.shields.io/badge/version-v2.0.2-blue.svg)](https://github.com/JesseBonanno/IndeterminateBeam/releases/tag/v2.0.2)
+[![Version](https://img.shields.io/badge/version-v2.0.3-blue.svg)](https://github.com/JesseBonanno/IndeterminateBeam/releases/tag/v2.0.3)
 [![License](https://img.shields.io/badge/license-MIT-lightgreen.svg)](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/LICENSE.txt)
 [![Documentation Status](https://readthedocs.org/projects/indeterminatebeam/badge/?version=main)](https://indeterminatebeam.readthedocs.io/en/main/?badge=main)
 [![Build Status](https://travis-ci.org/JesseBonanno/IndeterminateBeam.svg?branch=main)](https://travis-ci.org/JesseBonanno/IndeterminateBeam)
@@ -34,72 +34,85 @@ for educational purposes. The beambending module, although well documented, can 
 
 ## Functionality and Usage
 
-A typical use case of the `indeterminatebeam` package involves the following steps:
+A typical use case of the ```IndeterminateBeam``` package involves the following steps:
 
 1. Create a `Beam` object
 2. Create `Support` objects and assign to `Beam`
-3. Create `load` objects and assign to `Beam`
+3. Create `Load` objects and assign to `Beam`
 4. Solve for forces on `Beam` object
 5. Plot results
 
-Default units are kN, m, and kN.m   _(except for E (MPa), I (mm4), A(mm2))_
+You can follow along with the example below in this web-based [Jupyter Notebook](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.ipynb). Units and load direction conventions are described in the [package documentation](https://indeterminatebeam.readthedocs.io/en/main/theory.html).
 
-Load convention is described in the [package documentation](https://indeterminatebeam.readthedocs.io/en/main/).
+### Creating a Beam
 
-You can follow along with the example below in this web based notebook: [![Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.ipynb)
-
-##### Creating Beam
-
-The creation of a beam instance involves the input of the beam length (m) and optinally the input of the young's modulus (E), and second moment of area (I). E and I are optional and by default are the properties of a steel 150UB18.0. For a beam with constant EI (bending rigidity) these parameters will only affect the deflections calculated and not the distribution of forces.
+The creation of a `Beam` instance involves the input of the beam length (m) and optionally the input of the Young's Modulus (E), second moment of area (I), and cross-sectional area (A). E, I and A are optional and by default are the properties of a steel 150UB18.0. For a beam with constant properties, these parameters will only affect the deflections calculated and not the distribution of forces, unless spring supports are specified.
 
 ```python
 from indeterminatebeam import Beam
-beam = Beam(7)                          # Initialize a Beam object of length 5m with E and I as defaults
-beam_2 = Beam(9,E=2000, I =100000)      # Initialize a Beam object of length 9m with E and I assigned by user
+# Create 7 m beam with E, I, A as defaults
+beam = Beam(7)                          
+# Create 9 m beam with E, I, and A assigned by user
+beam_2 = Beam(9,E=2000, I =10**6, A = 3000)     
 ```
 
-##### Defining Supports
-Support objects are created separately from the beam object, and are defined by an x-coordinate (m) and the beams translational and rotational degrees of freedom.
+### Defining Supports
[…]
