### 2021-04-11 #111 @JesseBonanno
> Regarding the software, it is giving wrong results. Even the example in the README is wrong since it does not satisfy equilibrium in the vertical direction.

@nicoguaro I have fixed the issue and updated the package to version v2.0.3.

I have also slightly modified the readme example to use a "UDLV" class type rather than a "DistributedLoadV", since it is more relevant to the example. The original way that the code is written does work though with the new changes, but the DistributedLoad class is for arbritrary loads and runs the example slower than just using UDL class hence the change.

I plan on adding more tests to the unittesting, and plan on updating the package documentation to better reflect changes from moving to version 2 of the software, although i can do this in parallel to the JOSE review.
