### 2021-04-12 JesseBonanno/IndeterminateBeam#21 @nicoguaro
Problem with simply supported beam with end moment
I was testing the example in Figure 12-40 from Hibbeler 8th edition

![image](https://user-images.githubusercontent.com/1097787/114464345-3ffa1900-9bab-11eb-8f04-77e45dc98e8f.png)

But I obtain the following error message

```python
<lambdifygenerated-7>:2: RuntimeWarning:

invalid value encountered in multiply

<lambdifygenerated-7>:2: RuntimeWarning:

invalid value encountered in double_scalars

<lambdifygenerated-8>:2: RuntimeWarning:

invalid value encountered in multiply

<lambdifygenerated-9>:2: RuntimeWarning:

invalid value encountered in double_scalars

<lambdifygenerated-9>:2: RuntimeWarning:

invalid value encountered in multiply
```

with the following reactions

![image](https://user-images.githubusercontent.com/1097787/114464455-6750e600-9bab-11eb-84d5-fe2b9177ee75.png)

The code is the following

```python
from indeterminatebeam import Beam
from indeterminatebeam import Support
from indeterminatebeam import PointTorque, DistributedLoadV

beam1 = Beam(12)
beam1.add_loads(DistributedLoadV(-5, (0, 12)),
                PointTorque(20, 0),
                PointTorque(-20, 12))
beam1.add_supports(Support(0, (1, 1, 0)),
                   Support(12, (0, 1, 0)))
beam1.analyse()
```

If a remove the point moments, or add them in the interior the error disappears.
