### 2021-04-12 #111 @nicoguaro
@JesseBonanno, I installed the new version (this time in Linux) and it worked. The example seems to be working as well.

I created and [issue](JesseBonanno/IndeterminateBeam#21) for an example that does not work.
