### 2021-04-13 JesseBonanno/IndeterminateBeam#21 @JesseBonanno
Great observation, thank you for finding that issue. I have fixed it in the last commit.

![image](https://user-images.githubusercontent.com/53243847/114524608-b6207d80-9c88-11eb-8be4-7217f1139ca2.png)
