### 2021-04-13 #111 @JesseBonanno
@nicoguaro

> I installed the new version (this time in Linux) and it worked. The example seems to be working as well.

That's great, thanks.

> I created and [issue](JesseBonanno/IndeterminateBeam#21) for an example that does not work.

I have just fixed and closed the issue. Great spot, thank you for finding it.
