+++ tests/test_indeterminatebeam.py
@@ -1,8 +1,4 @@
-import sys
-import os
-sys.path.insert(0, os.path.abspath('../'))
-
-from indeterminatebeam.indeterminatebeam import (
+from indeterminatebeam import (
     Support, 
     Beam, 
     PointTorque,
