+++ indeterminatebeam/indeterminatebeam.py
@@ -28,12 +28,12 @@ from plotly.subplots import make_subplots
 import plotly.graph_objects as go
 
 # Local Application Imports
-from .data_validation import (
+from indeterminatebeam.data_validation import (
     assert_number,
     assert_positive_number,
     assert_strictly_positive_number, assert_length
 )
-from .loading import (
+from indeterminatebeam.loading import (
     PointLoad,
     PointLoadV,
     PointLoadH,
@@ -48,7 +48,7 @@ from .loading import (
     UDLH,
     UDLV
 )
-from .plotly_drawing_aid import (
+from indeterminatebeam.plotly_drawing_aid import (
     draw_line,
     draw_arrowhead,
     draw_arrow,
