+++ tests/test_indeterminatebeam.py
@@ -1,4 +1,8 @@
-from indeterminatebeam import (
+import sys
+import os
+sys.path.insert(0, os.path.abspath('../'))
+
+from indeterminatebeam.indeterminatebeam import (
     Support, 
     Beam, 
     PointTorque,
