+++ tests/test_indeterminatebeam.py
@@ -1,7 +1,3 @@
-import sys
-import os
-sys.path.insert(0, os.path.abspath('../'))
-
 from indeterminatebeam.indeterminatebeam import (
     Support, 
     Beam, 
