+++ .travis.yml
@@ -24,9 +24,9 @@ install:
 
 # command to test examples and then run tests
 script:
-  - pytest --cov= tests/
-  - pytest --cov= indeterminatebeam/
-  - pytest --cov= docs/examples/
+  - pytest --cov=tests/
+  - pytest --cov=indeterminatebeam/
+  - pytest --cov=docs/examples/
 
 # Push the results back to codecov
 after_success:
