### 2021-04-16 451b5e6
single point load error fix
Fixed issue #22 that occured when a single point load was added into the model. Issue originated from the sympy to piecewise conversion function.
