### 2021-04-16 JesseBonanno/IndeterminateBeam#22 @JesseBonanno
Issue concerned SingularityFunction to Piecewise conversion function.
