### 2021-04-16 JesseBonanno/IndeterminateBeam#21 @JesseBonanno
Yes, that sounds great
