### 2021-04-17 #111 @amcc1996
I have just finished my first thorough pass on the repository and paper. With regard to the paper itself, I have no further comments, besides the small remarks and corrections that @JesseBonanno has already applied. Concerning the overall code + repository, I have both some minor and major comments that I would like to discuss.

**Major revision comments**
1. I cannot run the [README.md example](https://gist.githubusercontent.com/amcc1996/5973b8cb62c8c845d5dce6d006107fa1/raw/841da117a7fefa8b7a806393f10dca11068a1d79/readme_example.py) in my system. I installed the package in a fresh virtual environment, so hopefully, everything is being run with the correct versions. After running the script, the following error surges in sympy:
```
Traceback (most recent call last):
  File "readme_example.py", line 38, in <module>
    beam.analyse()
  File "/home/amcc/repos/tmp/venv/lib/python3.8/site-packages/indeterminatebeam/indeterminatebeam.py", line 682, in analyse
    ans = float(ans)
  File "/home/amcc/repos/tmp/venv/lib/python3.8/site-packages/sympy/core/expr.py", line 358, in __float__
    raise TypeError("can't convert expression to float")
TypeError: can't convert expression to float
```
At the moment, I can't sort out whether the problem is in the code or on my specific setup. This error occurs sporadically in some other examples that I tried, but without any evident pattern. @JesseBonanno, can you please confirm that you can run the example in the link above?

2. I found the physical units adopted for the quantities involved rather odd. For instance, _meters_ are used for lengths, while _millimetre_ are used for the cross-sectional properties, loads are given in _kN_ and the Young modulus in _MPa_. Would it make more sense to use a consistent system of units? For instance, _m-N-Pa_ or _mm-N-MPa_. I must confess that I struggled a bit to compare some analytical solutions to the results of the code, as several corrections to the units are required. If this is current practice on the field, I completely understand the options, but otherwise, it might be a big source of confusion. What do you think @JesseBonanno ?

3. After running some benchmarks against analytical solutions, I have some doubts about the deflection computation. The code is not giving the correct deflection in some problems ([example](https://gist.githubusercontent.com/amcc1996/5973b8cb62c8c845d5dce6d006107fa1/raw/841da117a7fefa8b7a806393f10dca11068a1d79/console.py)). @JesseBonanno  please give a look at this example and the analytical solution for the deflection on the loading point. Maybe I am not seeing it right, but it does some that the computation is incorrect.

4 Another strange behaviour occurs with this [example](https://gist.githubusercontent.com/amcc1996/5973b8cb62c8c845d5dce6d006107fa1/raw/841da117a7fefa8b7a806393f10dca11068a1d79/pinned_beam.py). Apparently, the solver cannot deal with this setup, and gives some infinite reaction forces. Adding to my stupefaction, the online app can solve it seamlessly (even though the deflections seems wrong, see the analytical solution also provided in the script). Please try to run this script @JesseBonanno and assess if my comments are in line with your own observations.

**Minor revision comments**
1. In the meantime, the version changed from 2.0.2 to 2.0.3. Could you update it, @moorepants ? I don't know if I have the right permissions to do that.
2. The web application that runs the code is truly useful. However, I think there is no mention to it in the README.md page beyond the small badge on the top. Perhaps the repository could profit if there was some short mention in the *Functionality and usage* section. Any thought on this @JesseBonanno ?
3. The way I see it, the CONTRIBUTING.md file already encompasses all the community guidelines with respect to contributing, issue reporting and support. However, maybe it would be helpful to provide a *Support* section directly in the README.md file (or a SUPPORT.md file), where one could find your contact (email) and maybe the references already provided in the documentation. This is a minor suggestion, if you leave it as is, I will be good with it.
4. The code coverage badge does not make justice to the coverage reports with pytest-cov. On my computer, it registered about 75% coverage, instead of the poor 8% that are shown on the homepage!
5. Maybe the repository could also provide the example in the README.md file in a separate file. This would save the readers and users to copy and paste the code by hand to get the first nicely functioning example.
6. The aesthetics of some error messages is somewhat weird. For instance:
![indeterminatebeam-error](https://user-images.githubusercontent.com/43611757/115122181-85a15200-9fae-11eb-9ace-0e5798f419de.png)
This is not a problem at all. Anyway, I felt the need to report it, in the case this was not your original intention, @JesseBonanno .

I apologise for the long comments, which is exactly the opposite that is asked for in the guidelines. If you feel that some of this discussion can be moved to separate issues, please let me know @moorepants, and I will promptly migrate them.
