### 2021-04-18 82507c1
README.md update and minor fixes
Updated readme file and added a SUPPORT.md file which is referenced in the readme. Updated version to 2.0.4. Made minor change to how some raised errors are presented.
