### 2021-04-19 #111 @JesseBonanno
Hi @amcc1996,

First of all thank you for the time you have put into this. Your feedback is constructive and I appreciate it.

Second of all to make sure we are all on the same page, I have addressed issues here and previous issues found by @nicoguaro and have now published a version 2.0.4, which addresses all issues to date.

If we can move the review to this version (2.0.4) that would be best.

### Major revision comments

1.	I think this might be related to previous issues that I have addressed with @nicoguaro , as I recognise the source of this error message and it is something that he previously picked up on. I don’t recall seeing issues with that particular example and don’t have issues myself trying right now, so maybe you can re-run with v2.0.4 and see if the issue still occurs.
2.	I understand this sentiment, and initially considered having everything in SI units. I have instead tried to keep units aligned with the conventional units specified in my personal experience with engineering in the industry. The objective was both clarity and to avoid typing unnecessary 0’s if possible. I think the current set up is okay, but a third opinion would be good @nicoguaro
3.	I think the deflection is correct in the program as it currently is. I took a look at your example and I had a different formula for analytical deflection which had the same result as the calculator.
![image](https://user-images.githubusercontent.com/53243847/115310013-6d992080-a1b0-11eb-8515-f9f7586fa69d.png)

4.	As per issue 1, this is an issue that I believe was addressed [here](https://github.com/JesseBonanno/IndeterminateBeam/issues/22). The reason it works online is the code online is more up to date in line with v2.0.4.

### Minor Comments
1.	Lets change to v2.0.4
2.	Thanks, sounds like a good idea. I have reworded a small section of the README.md (the Statement of Need now Project Purpose) such that it better aligns with the current state of the project. There is a stronger emphasis on the existence of the web-page now.
3.	Makes sense to me. I have added in some references at the bottom of the readme and a small support section as suggested.
4.	Wow, 75% is a number that sounds more in line with reality than what I am getting. I have been having trouble getting the coverage to work properly and have been trying for longer than I think is worth it anymore. I am just going to remove it now.
5.	The example is currently in a separate file in docs/examples folder. I have linked a file for the ipynb file and the python file in the read me now at the start of the example.
6.	Yep, not my original intention for it to show like that. I was trying to avoid having the code line exceed too many characters. I have slightly modified the code, not sure if there is a better way to do this?

 Old code
![image](https://user-images.githubusercontent.com/53243847/115310146-af29cb80-a1b0-11eb-8e97-8d71bd97b042.png)

New code
![image](https://user-images.githubusercontent.com/53243847/115310188-be107e00-a1b0-11eb-97d3-9d0435136f13.png)
