### 2021-04-19 #111 @nicoguaro
I haven't move to the next step in my review, I was first checking the capabilities. Now I think that it is working properly and maybe next weekend will have some comments regarding documentation, testing and code.

Regarding the specific point on units, I have the same comment as @amcc1996 . At first I thought that it might be something related to Hibbeler's book, but there they use different systems of units (usually consistent). I consider that the use of consistent system of units is important. In the case of the software I think that it is even more important because it add [inessential weirdness](http://www.classmatters.org/2006_07/its-not-them.php) to the problem. You are adding a complexity layer that, I think, is not the point being discussed.
