### 2021-04-20 #111 @amcc1996
> 1. I think this might be related to previous issues that I have addressed with @nicoguaro , as I recognise the source of this error message and it is something that he previously picked up on. I don’t recall seeing issues with that particular example and don’t have issues myself trying right now, so maybe you can re-run with v2.0.4 and see if the issue still occurs.

Ok that would explain some of my observations. I was running version v2.0.3 on my system the last weekend. I will repeat my tests the next weekend with the new version, thank you for noticing that!

> 3. I think the deflection is correct in the program as it currently is. I took a look at your example and I had a different formula for an analytical deflection which had the same result as the calculator.

Well, it seems that we are using distinct analytical solutions. Perhaps I took the wrong one out of the shelf. I will take another look on my next pass and come back to you right after! Possibly, that is because I was looking at the deflection at the loading point and not the maximum deflection. Anyway, I will test it once again with the new version.

> 5. The example is currently in a separate file in docs/examples folder. I have linked a file for the ipynb file and the python file in the read me now at the start of the example.

My bad, maybe I didn't look hard enough.

> 6. Yep, not my original intention for it to show like that. I was trying to avoid having the code line exceed too many characters. I have slightly modified the code, not sure if there is a better way to do this?

That is great! A much better and ultimately pretty option, in my opinion! I think you can even remove the `+` operator, and Python will handle it seamlessly.
