### 2021-04-20 #111 @JesseBonanno
Based on the collective feedback I will update the package to use SI units. I am not sure when i will have time to do this but should be done by the weekend and i will make a comment here when it has been changed.

> I think you can even remove the + operator, and Python will handle it seamlessly.

Thanks for the tip, i will try it out!
