### 2021-04-20 #111 @moorepants
> If you feel that some of this discussion can be moved to separate issues, please let me know @moorepants, and I will promptly migrate them.

Discussion is fine here, but please open any issues on the repo that are needed.
