### 2021-04-25 #111 @amcc1996
@JesseBonanno, I repeated by evaluation with the up-to-date version `v2.0.4`. Unfortunately, some of my concerns persist. In the next points, I will summarise the observations resulting from this second pass.

1. I still cannot run the [readme_example.py](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.py) in my system. The error occurs in [line 685 of indeterminatebeam.py](https://github.com/JesseBonanno/IndeterminateBeam/blob/82507c1e1495a9f8b0eaadc2d89b63d168c3ea73/indeterminatebeam/indeterminatebeam.py#L685). If I print the value of `var` and `ans` at that point I get `y_0` and `-0.452244897959184*C1 - 0.0194285714285698`, respectively. To my knowledge, it seems that the error is a consequence of the symbolic variable `C1` that is somehow retained up to that point. I can replicate this error on my laptop with Python version `3.8.5`. However, when I run the same example with an older version (3.6, if I recall correctly), there is no error whatsoever. @JesseBonanno what version of Python are you using? Also, do you have any guess why is this behaving this way?

2. It seems that the `UDLV` class is missing from the imports statements of [example_1.py](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/example_1.py). Do you confirm @JesseBonanno ?

3. I can now confirm that the program returns the correct values of deflection in the benchmarks that I ran. The problem was on my side: in the pinned beam example, I was looking at the displacement at the wrong point; in the console beam example there was a sign-related issue in the analytical formula.

In sum, I think nearly all relevant points have been addressed and confirmed. The only exception is the README example which for some unknown reason exhibits unexpected behaviour. In truth, that error is not exclusive to this example, as it has appeared sporadically during my review. Any thoughs on this @JesseBonanno ?
