### 2021-04-25 #111 @nicoguaro
> @JesseBonanno, I repeated by evaluation with the up-to-date version `v2.0.4`. Unfortunately, some of my concerns persist. In the next points, I will summarise the observations resulting from this second pass.
>
>     1. I still cannot run the [readme_example.py](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.py) in my system. The error occurs in [line 685 of indeterminatebeam.py](https://github.com/JesseBonanno/IndeterminateBeam/blob/82507c1e1495a9f8b0eaadc2d89b63d168c3ea73/indeterminatebeam/indeterminatebeam.py#L685). If I print the value of `var` and `ans` at that point I get `y_0` and `-0.452244897959184*C1 - 0.0194285714285698`, respectively. To my knowledge, it seems that the error is a consequence of the symbolic variable `C1` that is somehow retained up to that point. I can replicate this error on my laptop with Python version `3.8.5`. However, when I run the same example with an older version (3.6, if I recall correctly), there is no error whatsoever. @JesseBonanno what version of Python are you using? Also, do you have any guess why is this behaving this way?

I am also having problems with that example.  I just installed last version and this is the output I get:

```
Traceback (most recent call last):
  File "/home/nicoguaro/Documents/reviews/readme_example.py", line 16, in <module>
    beam.analyse()
  File "/home/nicoguaro/Workspace/IndeterminateBeam/indeterminatebeam/indeterminatebeam.py", line 682, in analyse
    ans = float(ans)
  File "/home/nicoguaro/anaconda3/envs/indeterminate_beam/lib/python3.9/site-packages/sympy/core/expr.py", line 358, in __float__
    raise TypeError("can't convert expression to float")
TypeError: can't convert expression to float
```
