### 2021-04-27 #111 @JesseBonanno
@amcc1996 @nicoguaro
Sorry for the slow replies lately, i have been very overwhelmed with work lately and finding it hard to keep up at the moment.

1. In regards to point 1, i have tried running the examples on my computer again and did not observe the issue. I will have to try set up a fresh install to try see what you are seeing. I do have a vague idea of what the error means, it seems like their might be an equation missing that is needed to resolve the problem into a single solution. As to why there might be an equation missing i am not too sure.
2. Fixed, thanks.
3. Okay cheers.

On my end i will work to resolve the following issues:
1. Change units to SI
2. Determine why readme error occurs and fix.
