### 2021-04-28 #111 @moorepants
@amcc1996 and @nicoguaro thanks for the careful reviewing up till this point. It seems that @nicoguaro's checklist is mostly complete. @amcc1996 please update yours further if you haven't. I see there are some outstanding issues that @JesseBonanno is working on. @JesseBonanno, once you've made the updates please report back here. I'm going to be away from the internet till Monday, so I'll check on the status then.
