### 2021-05-01 #111 @amcc1996
@moorepants for now I have nothing to add besides the weirdly non-functioning example file in README.md. I believe this is the last point that is missing from my checklist. As soon as we discover what is going wrong with this, I will make just a quick check and get back to the discussion.
