+++ paper.md
@@ -84,10 +84,10 @@ The creation of a `Beam` instance involves the input of the beam length (m) and 
 
 ```python
 from indeterminatebeam import Beam
-# Create 7 m beam with E, I, A as defaults
-beam = Beam(7)                          
-# Create 9 m beam with E, I, and A assigned by user
-beam_2 = Beam(9,E=2000, I =10**6, A = 3000)     
+# Create 7000 mm beam with E, I, A as defaults
+beam = Beam(7000)                          
+# Create 9000 mm beam with E, I, and A assigned by user
+beam_2 = Beam(9000, E=2000, I =10**6, A = 3000)     
 ```
 
 ### Defining Supports
@@ -99,12 +99,12 @@ Optionally, stiffness can be specified in either of the translational directions
 
 ```python
 from indeterminatebeam import Support
-# Defines a pin support at location x = 5m  
-a = Support(5,(1,1,0))      
-# Defines a roller support at location x = 0m
-b = Support(0,(0,1,0))      
-# Defines a fixed support at location x = 7m
-c = Support(7,(1,1,1))      
+# Defines a pin support at location x = 5000 mm  
+a = Support(5000, (1,1,0))      
+# Defines a roller support at location x = 0 mm
+b = Support(0, (0,1,0))      
+# Defines a fixed support at location x = 7000 mm
+c = Support(7000, (1,1,1))      
 # Assign the support objects to a beam object created earlier
 beam.add_supports(a,b,c)    
 ```
@@ -114,12 +114,12 @@ beam.add_supports(a,b,c)
 
 ```python
 from indeterminatebeam import PointLoadV, PointTorque, DistributedLoadV
-# Create 1kN point load at x = 2m
-load_1 = PointLoadV(1,2)
-# Create a 2kN UDL from x = 1m to x = 4m
-load_2 = DistributedLoadV(2,(1,4))
-# Defines a 2kN.m point torque at x = 3.5m
-load_3 = PointTorque(2, 3.5)
+# Create a 1000 N point load at x = 2000 mm
+load_1 = PointLoadV(1000, 2000)
+# Create a 2 N/mm UDL from x = 1000 mm to x = 4000 mm
+load_2 = DistributedLoadV(2, (1000, 4000))
+# Defines a 2 kN.m point torque at x = 3500 mm
+load_3 = PointTorque(2*10**6, 3500)
 # Assign the load objects to the beam object
 beam.add_loads(load_1,load_2,load_3)
 ```
