+++ indeterminatebeam/indeterminatebeam.py
@@ -3,17 +3,16 @@ and auxillary class for Support.
 
 Example
 --------
->>> beam = Beam(6)
+>>> beam = Beam(6000)
 >>> a = Support()
->>> c = Support(6,(0,1,0))
+>>> c = Support(6000,(0,1,0))
 >>> beam.add_supports(a,c)
->>> beam.add_loads(PointLoadV(-15,3))
+>>> beam.add_loads(PointLoadV(-150,3000))
 >>> beam.analyse()
 >>> beam.plot_beam_external()
 >>> beam.plot_beam_internal()
 
 """
-
 
 # Standard Library Imports
 from collections import namedtuple
@@ -73,30 +72,30 @@ class Support:
     Attributes:
     -------------
         _position: float
-            x coordinate of support on a beam (default 0)
+            x coordinate of support on a beam (mm) (default 0)
         _stiffness: tuple of 3 floats or infinity
-            stiffness K (kN/mm) for movement in x, y and bending, oo
+            stiffness K (N/mm) for movement in x, y and bending. oo
             represents infinity in sympy and means a completely fixed
             conventional support, and 0 means free to move.
         _DOF : tuple of 3 booleans
             Degrees of freedom that are restraint on a beam for
-            movement in x, y and bending, 1 represents that a reaction
+            movement in x, y and bending. 1 represents that a reaction
             force exists and 0 represents free (default (1,1,1))
         _fixed: tuple of 3 booleans
             Degrees of freedom that are completely fixed on a beam for
-            movement in x, y and bending, 1 represents fixed and 0
+            movement in x, y and bending. 1 represents fixed and 0
             represents free or spring  (default (1,1,1))
 
     Examples
     --------
     >>> # Creates a fixed suppot at location 0
     >>> Support(0, (1,1,1))
-    >>> # Creates a pinned support at location 5
-    >>> Support(5, (1,1,0))
-    >>> # Creates a roller support at location 5.54
-    >>> Support(5.54, (0,1,0))
-    >>> # Creates a y direction spring support at location 7.5
-    >>> Support(7.5, (0,1,0), ky = 5)
+    >>> # Creates a pinned support at location 5000 mm
+    >>> Support(5000, (1,1,0))
+    >>> # Creates a roller support at location 5540 mm
+    >>> Support(5540, (0,1,0))
+    >>> # Creates a y direction spring support at location 7500 mm
+    >>> Support(7500, (0,1,0), ky = 5000)
     """
 
     def __init__(self, coord=0, fixed=(1, 1, 1), kx=None, ky=None):
@@ -106,18 +105,18 @@ class Support:
         Parameters:
         -----------
         coord: float
-            x coordinate of support on a beam (default 0)
+            x coordinate of support on a beam (mm) (default 0)
             (default not 0.0 due to a float precision error that
             previously occured)
         fixed: tuple of 3 booleans
             Degrees of freedom that are fixed on a beam for movement in
-            x, y and bending, 1 represents fixed and 0 represents free
+            x, y and bending. 1 represents fixed and 0 represents free
             (default (1,1,1))
         kx :
-            stiffness of x support (kN/mm), if set will overide the
+            stiffness of x support (N/mm), if set will overide the
             value placed in the fixed tuple. (default = None)
         ky : (positive number)
-            stiffness of y support (kN/mm), if set will overide the
+            stiffness of y support (N/mm), if set will overide the
             value placed in the fixed tuple. (default = None)
         """
         # validate coordinate
@@ -182,7 +181,7 @@ class Beam:
     _x0 :float
         Left end coordinate of beam (always defined as 0).
     _x1 :float
-        Right end coordinate of beam.
+        Right end coordinate of beam (in mm).
 
[…]
