### 2021-05-02 ba8ed6d
Update docstrings.rst
