### 2021-05-02 #111 @JesseBonanno
Hi all,

Finally managed to get a good chunk of free time this sunday afternoon and have managed to adress all our issues to date:

1. Changed all references to refer to base SI units.  Changes occured for several python files, all examples, unit tests, example png snapshots, readme.md and of course paper.md . As there is a lot of references to units throught the code i wouldnt be surprised if i missed any small references to the old units although i have tried my best to cover everything.

2. Readme error. This one took a while to figure out but it seems to be a float precision error with linsolve which only happens with Sympy version 1.8 which was released 22 days ago. I tested on sympy 1.4 which is the minimum version i specified in the requirements and with only that change alone the code worked. I also tested on version 1.7.1 which is the version before 1.8 and the version i personally have installed at the moment and it also worked. I think the easiest solution at the moment would be to have sympy==1.7.1 in the requirements, but i recall reading that forcing a particular version is bad practice since it could make someone down grade in future. Please advise if this solution is adequate.

3. As per an issue created by @nicoguaro i have updated documentation to include all load classes.
