+++ indeterminatebeam/loading.py
@@ -62,15 +62,15 @@ class PointTorque(Load):
     Parameters:
     -----------
     force: float
-        Torque in N.mm
+        Torque in kN.m
         Note: named force for simplicity.
     coord: float
-        x coordinate of torque on beam in mm
+        x coordinate of torque on beam
 
     Examples
     ---------
     >>> # 30 kN·m (clockwise) torque at x=4 m
-    >>> motor_torque = PointTorque(30*(10**6), 4)
+    >>> motor_torque = PointTorque(30, 4)
 
     """
 
@@ -104,9 +104,9 @@ class PointLoad(Load):
     Parameters:
     -----------
     Force: float
-        Force in Newtons (N)
+        Force in kN
     coord: float
-        x coordinate of load on beam in mm
+        x coordinate of load on beam
     angle: float
         angle of point load where:
         - 0 degrees is purely horizontal +ve
@@ -116,12 +116,13 @@ class PointLoad(Load):
 
     Examples
     --------
-    >>> # 100 N towards the right at x=9000 mm
-    >>> external_force = PointLoad(100, 9000, 90)
-    >>> # 300 N downwards at x=3000 mm
-    >>> external_force = PointLoad(-300, 3000, 0)
+    >>> # 10 kN towards the right at x=9 m
+    >>> external_force = PointLoad(10, 9, 90)
+    >>> # 30 kN downwards at x=3 m
+    >>> external_force = PointLoad(-30, 3, 0)
     >>> external_force
-        PointLoad(force=-300, coord=3000, angle=0)
+        PointLoad(force=-30, coord=3, angle=0)
+    
     """
 
     def __init__(self, force=0, coord=0, angle=0):
@@ -153,7 +154,7 @@ class UDL(Load):
     Parameters
     ----------
     force : int, optional
-        UDL load in N/mm, by default 0
+        UDL load in kN/m, by default 0
     span: tuple of floats
         A tuple containing the starting and ending coordinate that
         the UDL is applied to.
@@ -165,8 +166,8 @@ class UDL(Load):
 
     Examples
     --------
-    >>> # load of 1 N/mm from 1000 <= x <= 4000 mm (vertical)
-    >>> self_weight = UDL(1, (1000, 4000), 90)
+    >>> # load of 1 kN/m from 1 <= x <= 4 (vertical)
+    >>> self_weight = UDL(1, (1, 4), 90)
     """
 
     def __init__(self, force=0, span=(0, 0), angle=0):
@@ -215,7 +216,7 @@ class TrapezoidalLoad(Load):
     ----------
     force : tuple of floats
         A tuple containing the starting and ending loads of
-        the trapezoidal load in N/mm.
+        the trapezoidal load in kN/m.
     span: tuple of floats
         A tuple containing the starting and ending coordinate that
         the trapezoidal load is applied to.
@@ -227,9 +228,9 @@ class TrapezoidalLoad(Load):
 
     Examples
     --------
-    >>> # trapezoidal load starting at 2 N/mm at 1000 mm and ending at 3 N/mm
-    >>> # at 4000 mm (vertical)
-    >>> self_weight = UDL((2,3), (1000, 4000), 90)
+    >>> # trapezoidal load starting at 2 kN/m at 1 m and ending at 3 kN/m
+    >>> # at 4 m (vertical)
+    >>> self_weight = UDL((2,3), (1, 4), 90)
     """
 
     def __init__(self, force=(0, 0), span=(0, 0), angle=0):
@@ -318,8 +319,8 @@ class DistributedLoad(Load):
         - 180 degrees is purely horizontal -ve of force sign specified.
     Examples
     --------
[…]
