+++ app.py
@@ -1557,4 +1557,4 @@ def report(n, graph_1, graph_2, results, json):
 
 
 if __name__ == '__main__':
-    app.run_server(debug=True)
+    app.run_server()
