### 2021-05-03 #111 @moorepants
> This one took a while to figure out but it seems to be a float precision error with linsolve which only happens with Sympy version 1.8 which was released 22 days ago.

Is this related: https://github.com/sympy/sympy/issues/21162
