### 2021-05-03 #111 @JesseBonanno
@moorepants Seems to be the exact issue. If a fix is released for that issue than I believe it would solve the issue observed here.
