### 2021-05-03 #111 @nicoguaro
@moorepants, I think that the package is ready from my perspective.

I have some minor comments that I think that can improve the package but they don't need to be addressed for me to accept:

- I would suggest to create a separate folder for the paper in the repository. The package could grow and the files related to the paper might be a distraction there.
- In the API (docstrings) for vertical and horizontal loads there isn't a note about what direction is positive and negative. I know that this is covered in the Theory section of the documentation but adding it there would cause no harm and might be useful for calls to `help()` or `?` in Jupyter.
- At the beginning of Theory in docs a diagram would make clearer the balance of forces and moments.
- I think that there is an extra `beambending` at the beginning of the Theory section.
- Consider adding more cases to the test suite. Maybe some of the cases raised during this review process.
