### 2021-05-04 94da71e
Minor updates
New directory for storing paper related documents, update to example 1 in documentation, update to graph size for plot_beam_diagram() and plot_reaction_forces()
