### 2021-05-04 #111 @JesseBonanno
@nicoguaro I have addressed your latest comments below.
> I would suggest to create a separate folder for the paper in the repository. The package could grow and the files related to the paper might be a distraction there.

I agree and have updated, will need to update repository to https://github.com/JesseBonanno/IndeterminateBeam/paper for whedon to work. If this is an issue can move paper back.

> In the API (docstrings) for vertical and horizontal loads there isn't a note about what direction is positive and negative. I know that this is covered in the Theory section of the documentation but adding it there would cause no harm and might be useful for calls to help() or ? in Jupyter.

Updated.

> At the beginning of Theory in docs a diagram would make clearer the balance of forces and moments.

Noted, will leave as is for the moment.

> I think that there is an extra beambending at the beginning of the Theory section.

Fixed.

> Consider adding more cases to the test suite. Maybe some of the cases raised during this review process.

Agree. I want to add all the examples to the test suite as they cover a broad range of cases that the package covers. I will work on this when I can.
