### 2021-05-04 #111 @moorepants
> @moorepants Seems to be the exact issue. If a fix is released for that issue than I believe it would solve the issue observed here.

If you can add a reproducible example to that issue, it could help.
