### 2021-05-04 #111 @moorepants
> I think the easiest solution at the moment would be to have sympy==1.7.1 in the requirements, but i recall reading that forcing a particular version is bad practice since it could make someone down grade in future. Please advise if this solution is adequate.

This is an appropriate solution. Or you could do `>=1.4,<1.8` to be more flexible.
