### 2021-05-04 #111 @moorepants
@amcc1996 there is one checkbox for the "examples" that you've left unchecked. Is that now resolved?
