### 2021-05-04 #111 @JesseBonanno
> > @moorepants Seems to be the exact issue. If a fix is released for that issue than I believe it would solve the issue observed here.
>
> If you can add a reproducible example to that issue, it could help.

The readme example works as a reproducible example, is this adequate?

> This is an appropriate solution. Or you could do >=1.4,<1.8 to be more flexible.

Noted, will update.

> @JesseBonanno Please make a new software release with the final changes.

Will do.
