### 2021-05-04 #111 @moorepants
> The readme example works as a reproducible example, is this adequate?

It can, but maybe a simple linsolve example with the values that makes it fail is best.
