+++ indeterminatebeam/version.py
@@ -1,1 +1,1 @@
-__version__ = 'v2.0.4'
+__version__ = 'v2.0.5'
