### 2021-05-04 #111 @JesseBonanno
> It can, but maybe a simple linsolve example with the values that makes it fail is best.

Understood, I will get something back to you soon. I will try to find a simple set of equations that raises the error.
