### 2021-05-04 #111 @JesseBonanno
@moorepants

Package has been updated to v2.0.5.

Example of Sympy v1.8 linsolve error occurs is for a 7000 mm simply supported beam with a 1000 N point load at its centre.

equations = [
y_0 + y_7000 + 1000.0,
7000*y_7000 + 3500000.0,
0.552486187845304*C2,
3867.40331491713*C1 + 0.552486187845304*C2 + 31583793738.4899*y_0 + 3947974217311.23
]
unknowns = [y_0, y_7000, C1, C2]

solution = (1102922357.72339*C1 - 3.37769972052787e+18, -500.0, 1.0*C1, 0)

Interestingly for the same problem (a simply supported beam with a point load at is centre) did not occur for a 2000, 3000, 4000, 5000 , 6000, 8000, 9000, 7, 70, 700 or 70,000 mm long beam which was every other case tested. A summary of the equations and unknowns in the case for the 2000 mm beam is presented below for comparison.

equations = [
y_0 + y_2000 + 1000.0,
2000*y_2000 + 1000000.0,
0.552486187845304*C2,
1104.97237569061*C1 + 0.552486187845304*C2 + 736648250.460405*y_0 + 92081031307.5506
]

unknown = [y_0, y_2000, C1, C2]

solution = (-500.0, -500.0, 250000000.0, 0)

Also worth noting that the example in which we observed this error first the beam length was also 7000 mm.

Also worth noting that this error does not occur if the Sympy installation is changed to 1.7.1 or 1.4 (other versions have not been tested).
