### 2021-05-07 #111 @amcc1996
@JesseBonanno the README example now works just fine!
@moorepants , with this I complete my checkbox list. However, I would really appreciate running some examples with the fully corrected version that I believe @JesseBonanno will release soon. Just to make one last check, since some modifications have been introduced relative to my last review.
