### 2021-05-08 #111 @moorepants
> However, I would really appreciate running some examples with the fully corrected version that I believe @JesseBonanno will release soon.

Sounds good.

@JesseBonanno, you have all current comments. Please make final adjustments and the reviewers will make a final pass after that.
