### 2021-05-12 af8c418
Unit update
Update units to be in default SI and added feature to allow changing the units that are used for analysis and presentation of results. Added a file called units that stores different units and their relation to the SI units. Updated all examples, paper.md, readme.md, example print outs and updated documentation to match the change to the program. Updated testing to include readme example with SI units and also with modified metric units.
