+++ paper/paper.md
@@ -76,7 +76,11 @@ A typical use case of the ```IndeterminateBeam``` package involves the following
 4. Solve for forces on `Beam` object
 5. Plot results
 
-You can follow along with the example below in this web-based [Jupyter Notebook](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.ipynb). Units and load direction conventions are described in the [package documentation](https://indeterminatebeam.readthedocs.io/en/main/theory.html).
+You can follow along with the example below in this web-based [Jupyter Notebook](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/readme_example.ipynb). 
+Alternatively, you can download the jupyter-notebook for this example [here](https://raw.githubusercontent.com/JesseBonanno/IndeterminateBeam/main/docs/examples/readme_example.ipynb), or the python file for this code [here](https://raw.githubusercontent.com/JesseBonanno/IndeterminateBeam/main/docs/examples/readme_example.py).
+
+Units and load direction conventions are described in the [package documentation](https://indeterminatebeam.readthedocs.io/en/main/theory.html).
+
 
 ### Creating a Beam
 
@@ -84,10 +88,10 @@ The creation of a `Beam` instance involves the input of the beam length (m) and 
 
 ```python
 from indeterminatebeam import Beam
-# Create 7000 mm beam with E, I, A as defaults
-beam = Beam(7000)                          
-# Create 9000 mm beam with E, I, and A assigned by user
-beam_2 = Beam(9000, E=2000, I =10**6, A = 3000)     
+# Create 7 m beam with E, I, A as defaults
+beam = Beam(7)                          
+# Create 9 m beam with E, I, and A assigned by user
+beam_2 = Beam(9, E=2000, I =10**6, A = 3000)     
 ```
 
 ### Defining Supports
@@ -99,12 +103,12 @@ Optionally, stiffness can be specified in either of the translational directions
 
 ```python
 from indeterminatebeam import Support
-# Defines a pin support at location x = 5000 mm  
-a = Support(5000, (1,1,0))      
-# Defines a roller support at location x = 0 mm
+# Defines a pin support at location x = 5 m  
+a = Support(5, (1,1,0))      
+# Defines a roller support at location x = 0 m
 b = Support(0, (0,1,0))      
-# Defines a fixed support at location x = 7000 mm
-c = Support(7000, (1,1,1))      
+# Defines a fixed support at location x = 7 m
+c = Support(7, (1,1,1))      
 # Assign the support objects to a beam object created earlier
 beam.add_supports(a,b,c)    
 ```
@@ -114,12 +118,12 @@ beam.add_supports(a,b,c)
 
 ```python
 from indeterminatebeam import PointLoadV, PointTorque, DistributedLoadV
-# Create a 1000 N point load at x = 2000 mm
-load_1 = PointLoadV(1000, 2000)
-# Create a 2 N/mm UDL from x = 1000 mm to x = 4000 mm
-load_2 = DistributedLoadV(2, (1000, 4000))
-# Defines a 2 kN.m point torque at x = 3500 mm
-load_3 = PointTorque(2*10**6, 3500)
+# Create a 1000 N point load at x = 2 m
+load_1 = PointLoadV(1000, 2)
+# Create a 2000 N/m UDL from x = 1 m to x = 4 m
+load_2 = DistributedLoadV(2000, (1, 4))
+# Defines a 2 kN.m point torque at x = 3.5 m
+load_3 = PointTorque(2*10**3, 3.5)
 # Assign the load objects to the beam object
 beam.add_loads(load_1,load_2,load_3)
 ```
