+++ indeterminatebeam/__init__.py
@@ -15,3 +15,4 @@ from indeterminatebeam.loading import (
     UDLV
 )
 from indeterminatebeam.data_validation import assert_length
+from indeterminatebeam.units import METRIC_UNITS, IMPERIAL_UNITS
