+++ indeterminatebeam/version.py
@@ -1,1 +1,1 @@
-__version__ = 'v2.0.5'
+__version__ = 'v2.1.0'

[…]
