### 2021-05-12 #111 @JesseBonanno
Hi all,

I have just finished making changes regarding the units. I decided to make the implementation a little bit more flexible since I would prefer to work in alternate units and so the package now uses SI units (N and m) by default but the units used can be changed using the update_units() method. The units available are shown in the units.py file that I have added to the indeterminatebeam folder.

I have updated all examples, example plot png images, paper.md file, readme.md file, testing file, docstrings and read the docs pages to match these changes. Graphs now also explicitly refer to units everywhere, which previously wasn't the case, to remove ambiguity that might be caused by this additional functionality.

The testing file also now includes the readme example and another interpretation of the readme example using different units to provide testing for the update_units() function and method.

I have yet to complete updating the web app to work with the update_units() method but I think it is irrelevant for our purposes, so I can work on it in parallel to the review here.

The update has been released as version v2.1.0.

The changes made in this version are fairly broad and i have made changes to the analysis function so there is certainly potential for error. I have double checked and walked over all the lines in the files myself and everything seems in good order to me. Although this was initially meant to be a quick double check on your behalf I would advise care in review with a moderate review of the examples in the simple_demo.ipynb file and any other examples you like to use.

Thanks in advance!
