### 2021-05-13 #111 @nicoguaro
@JesseBonanno, I think that I will need some time re-review since there is a lot of changes.

Regarding some quick review I did, I have the following comments:

- I think that there a couple of tutorials regarding the update of units and query points are needed in the documentation. For this type of functionality an API entry is not enough, in my opinion.
- The new web app is less intuitive than the one before.
- I found a mistake in the deflection for the query points. I'll open an issue.
