### 2021-05-13 #111 @JesseBonanno
@nicoguaro
`I think that I will need some time re-review since there is a lot of changes.`
Understandable, I agree.

`I think that a couple of tutorials regarding the update of units and query points are needed in the documentation. For this type of functionality an API entry is not enough, in my opinion.`
Okay that is a fair consideration, I currently only have one example in the documentation and will add others.

`The new web app is less intuitive than the one before.`
Currently in the process of fixing up the web app, when I am done I hope that it will appear very similiar to before except with the additional settings to play with units.

`I found a mistake in the deflection for the query points. I'll open an issue.`
This is a rounding issue where the program rounds to 3 dp, which was more so for presentation purposes. I can change the rounding to only occur for plotting purposes.

I also found an issue with Imperial Units not working which I have currently addressed with a few additional lines although I have not uploaded the solution yet.
