### 2021-05-16 a999797
Update to Web-App
Updated web app to work for unit changes. Fixed minor bug that prevented analysis for imperial units. Upgrade documentation to include more examples as taken from the simple_demo.ipynb example file. Other minor updates to reflect changes.
