+++ app.py
@@ -20,7 +20,15 @@ from indeterminatebeam.loading import (
 from datetime import datetime
 import time
 from indeterminatebeam.version import __version__
-from indeterminatebeam.units import IMPERIAL_UNITS, METRIC_UNITS
+from indeterminatebeam.units import IMPERIAL_UNITS, METRIC_UNITS, UNIT_KEYS, UNIT_VALUES
+from indeterminatebeam.data_validation import (
+    assert_number,
+    assert_positive_number,
+    assert_strictly_positive_number,
+    assert_length,
+    assert_list_contents,
+    assert_contents,
+)
 
 from dash_extensions import Download
 from dash.exceptions import PreventUpdate
@@ -110,23 +118,23 @@ copyright_ = dbc.Row(
 
 beam_table_data = {
     'Length': {
-        'init': 5000,
-        'units': ' mm',
+        'init': 5,
+        'units': ' m',
         'type': 'numeric'
     },
-    "Young's Modulus (MPa)": {
-        'init': 2 * 10**5,
-        'units': ' MPa',
+    "Young's Modulus": {
+        'init': 200 * 10**9,
+        'units': ' Pa',
         'type': 'numeric'
     },
-    "Second Moment of Area (mm4)": {
-        'init': 9.05 * 10**6,
-        'units': ' mm4',
+    "Second Moment of Area": {
+        'init': 9.05 * 10**-6,
+        'units': ' m4',
         'type': 'numeric'
     },
-    "Cross-Sectional Area (mm2)": {
-        'init': 2300,
-        'units': ' mm2',
+    "Cross-Sectional Area": {
+        'init': 0.23,
+        'units': ' m2',
         'type': 'numeric'
     },
 }
@@ -184,24 +192,24 @@ beam_content = dbc.Card(
 # R - Restraint, F- Free, or number for spring, Spring not an option for m.
 
 support_table_data = {
-    'Coordinate (mm)': {
+    'Coordinate': {
         'init': 0,
-        'units': ' mm',
+        'units': ' m',
         'type': 'numeric'
     },
     "X": {
         'init': 'R',
-        'units': ' N/mm',
+        'units': ' N/m',
         'type': 'any'
     },
     "Y": {
         'init': 'R',
-        'units': ' N/mm',
+        'units': ' N/m',
         'type': 'any'
     },
     "M": {
         'init': 'R',
-        'units': ' N/mm',
+        'units': ' N/m',
         'type': 'any',
     }
 }
@@ -234,7 +242,7 @@ support_instructions = dcc.Markdown('''
             2. For each direction specify one of the following:
                * f or F - Indicates a free support
                * r or R - Indicates a rigid support
-               * n - Indicates a spring stiffness of n N/mm
+               * n - Indicates a stiffness of n (default unit N/m)
                  (where n is (generally) a positive number)
 
             ''')
@@ -260,15 +268,15 @@ support_content = dbc.Card(
 # Basic support
 
 basic_support_table_data = {
-    'Coordinate (mm)': {
+    'Coordinate': {
         'init': 0,
         'type': 'numeric',
-        'presentation': 'input'
+        'presentation': 'input',
     },
     "Support": {
         'init': 'Fixed',
         'type': 'any',
-        'presentation': 'dropdown'
+        'presentation': 'dropdown',
     }
 }
 
@@ -338,12 +346,12 @@ basic_support_content = dbc.Card(
 # Properties for point_load Tab
 
 point_load_table_data = {
-    'Coordinate (mm)': {
+    'Coordinate': {
         'init': 0,
-        'units': ' mm',
+        'units': ' m',
         'type': 'numeric'
     },
-    "Force (N)": {
+    "Force": {
         'init': 0,
[…]
