### 2021-05-16 #111 @JesseBonanno
@nicoguaro I have now updated the web-app to work properly based on the update_unit() method. I have also updated the examples to cover all the examples shown in the jupyter notebook example.

I am not sure how to approach the rounding issue. I have changed the implementation slightly to make sure no rounding occurs on the backend. Numbers are all displayed using the f string :.3f method for display purposes. To get more precise numbers the user would have to either use the get_deflection() method which would provide no rounding or they would have to change the units to something more sensitive such as mm.
