### 2021-05-17 dc987f6
Update app.py
Added option to change default support type. Added unit to basic support table for coordinate.
