### 2021-05-17 #111 @moorepants
It seems @JesseBonanno has made changes to address all issues, including making some larger changes in the software as a byproduct of fixing the review comments. We'll need a final pass from both reviewers to ensure the checklists are still valid, as it sounds like some new issues could have been introduced. Is another two weeks a suitable for deadline for the final pass?
