+++ app.py
@@ -80,9 +80,6 @@ For more, you can view the following:
 * [![Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/JesseBonanno/IndeterminateBeam/blob/main/docs/examples/simple_demo.ipynb)
    The Python based Jupyter Notebook examples
 * [![Article](https://img.shields.io/badge/Article-Complete-green.svg)](https://github.com/JesseBonanno/IndeterminateBeam/blob/main/IndeterminateBeam_Article.pdf)
-
-Note: As the Python package calculations are purely analytical calculation \
-times can be relatively slow.
 ''')
 
 # the content for the sidebar
@@ -114,6 +111,62 @@ copyright_ = dbc.Row(
     ])
 
 
+
+def create_table(id_, table, init, row_deletable = True):
+    if init:
+        data = [init]
+    # used to initialise no data for the query table
+    else:
+        data = []
+
+    table = dash_table.DataTable(
+        id=id_,
+        columns=[{
+            'name': d,
+            'id': d,
+            'deletable': False,
+            'renamable': False,
+            'type': table[d]['type'],
+            'format': Format(
+                symbol=Symbol.yes,
+                symbol_suffix=table[d]['units'])
+        } for d in table.keys()],
+        data=data,
+        editable=True,
+        row_deletable=row_deletable,
+    )
+    return table
+
+
+def create_content(instruction_name, instructions, table, button_label="", button_id="", add_button = True):
+    if add_button:
+        _ = [
+            table,
+            html.Br(),
+            html.Button(button_label, id=button_id, n_clicks=0),
+            dbc.Collapse(
+                dbc.Card(dbc.CardBody(instructions)),
+                id=instruction_name,
+            ),
+        ]
+    else:
+        _ = [
+            table,
+            html.Br(),
+            dbc.Collapse(
+                dbc.Card(dbc.CardBody(instructions)),
+                id=instruction_name,
+            ),
+        ]
+
+    card = dbc.Card(
+        dbc.CardBody(_),
+        className="mt-3",
+    )
+
+    return card
+
+            
 # Properties for Beam Tab
 
 beam_table_data = {
@@ -141,22 +194,7 @@ beam_table_data = {
 
 beam_table_init = {k: v['init'] for k, v in beam_table_data.items()}
 
-beam_table = dash_table.DataTable(
-    id='beam-table',
-    columns=[{
-        'name': d,
-        'id': d,
-        'deletable': False,
-        'renamable': False,
-        'type': beam_table_data[d]['type'],
-        'format': Format(
-            symbol=Symbol.yes,
-            symbol_suffix=beam_table_data[d]['units'])
-    } for d in beam_table_data.keys()],
-    data=[beam_table_init],
-    editable=True,
-    row_deletable=False,
-)
+beam_table = create_table('beam-table', beam_table_data, beam_table_init, row_deletable=False)
 
 beam_instructions = dcc.Markdown('''
 
@@ -173,19 +211,8 @@ beam_instructions = dcc.Markdown('''
             is specified E and A will affect the load distribution for the horizontal loads only.
             ''')
 
-beam_content = dbc.Card(
-    dbc.CardBody(
-        [
-            beam_table,
-            html.Br(),
-            dbc.Collapse(
-                dbc.Card(dbc.CardBody(beam_instructions)),
-                id="beam_instructions",
-            ),
-        ]
-    ),
-    className="mt-3",
-)
+beam_content = create_content('beam_instructions', beam_instructions, beam_table)
+
 
 # Properties for (Advanced) Support Tab
 # Just do as a table but let inputs be
@@ -217,22 +244,7 @@ support_table_data = {
 
[…]
