### 2021-05-22 #111 @amcc1996
@JesseBonanno , I repeated the review process on the updated version (v2.1.0) and did not see a false note anywhere. The possibility to select proper physical units (with the SI system being set as default) is a massive improvement, in my opinion. The documentation was also taken to another level, and the sections pertaining to the treatment of the units is very clear. I ran the unit tests packages with the software and everything was fine. I also looked at all the examples I had created for the first pass and adapted them to the new unit framework and everything looks good!

The only recommendation I have at this moment is to update the paper to refer to the current implementation of the physical units (I believe it still uses mm in the exemplary code snippets in the last version). At first, I was also tempted to advise you to make some comments on the default physical units in the article, as well as the flexibility that the program entails in this regard, but now I am not really sure, as you already refer the reader to the online documentation. Perhaps @nicoguaro could also give some opinion on this.
