### 2021-05-23 2bf3d3f
Update to v2.1.2
Fixed minor bug for point torque at end of cantilever. Refactored web app. Updated paper.md and readme.md to mention unit update. Updated to version v2.1.2.
