+++ indeterminatebeam/version.py
@@ -1,1 +1,1 @@
-__version__ = 'v2.1.2'
+__version__ = 'v2.1.4'
