### 2021-05-23 #111 @JesseBonanno
@amcc1996 Glad to hear you see the value in the unit update, and thanks for the positive feedback.

I updated the paper.md and README.md to have a simple sentence mentioning the change as below:

> The units used throughout the Python package are the base SI units (newtons and metres), but can be updated using the update_units method. Units and load direction conventions are described in the package documentation.

I fixed a minor bug that was found earlier in the review process that I accidentally reintroduced in v2.1.0.

I also played around with the setup.py file a bit in order to get PyPi to show the readme.md as the description. As a result the version has now jumped ahead to v2.1.4.
