### 2021-05-27 #111 @moorepants
Ok, great! Thanks for the reminder. I see that all check boxes are done and it seems that all points have been addressed. I need to take a pass through the paper to make sure things are in order for the final publication and then I think we can approve. I'll try to do that tomorrow afternoon.
