### 2021-05-28 #111 @moorepants
@JesseBonanno I have checked over everything and it seems your paper can be accepted! Very nice work. You will need to archive your software (e.g. on Zenodo) for the version that will be tied to the paper (currently v2.1.4). Post the archieve DOI here. Once you make the archive I will ask the editor in chief to formally accept and publish the paper.

@nicoguaro and @amcc1996 thank you both very much for the excellent review. This package has many nice improvements from your comments. We greatly appreciate your time and energy.
