### 2021-05-29 #111 @JesseBonanno
> I have checked over everything and it seems your paper can be accepted!

Thats great! Thanks everyone for dedicating their time and knowledge in this review. Everyone made valuable contributions and put in a great deal of effort. I appreciate it very much. @nicoguaro @amcc1996  @moorepants

> You will need to archive your software (e.g. on Zenodo) for the version that will be tied to the paper (currently v2.1.4).

I had some difficulty with Zenodo generating the upload automatically with new Github releases (v2.1.5, v2.1.6). I have manually uploaded v2.1.6 via Zenodo and generated DOI **10.5281/zenodo.4840592**. There are no real changes between 2.1.4 and 2.1.6 that need to be reviewed, just updates to the version reference.
