### 2021-05-29 #111 @JesseBonanno
@moorepants yeah there does seem to be something wrong. I have contacted support and hopefully they can help to resolve the issue. The upload page [here](https://zenodo.org/record/4840592/#.YLI-K6gzY2w) seems to show everything in working order and has the DOI, but the DOI does not seem to be activated.

In the mean time i will try an alternative to zenodo.
