### 2021-05-30 #111 @JesseBonanno
@moorepants Zenodo is working now with the automatic Github web hook having just been properly processed, the DOI can be changed to 10.5281/zenodo.4851065 for version v2.1.6. The following URL works for me https://doi.org/10.5281/zenodo.4851065

![image](https://user-images.githubusercontent.com/53243847/120097726-831d4500-c175-11eb-957f-3c5adc5adace.png)
