### 2021-05-30 #111 @amcc1996
Congratulations on your publication @JesseBonanno! Your contribution to open source software and education is very valuable, and I truly believe that both students and engineers can find value in this work.
Thanks for the help and collaboration in the review process @moorepants and @nicoguaro!
