### 2021-05-31 #111 @labarba
hi @JesseBonanno 👋  — we do like to request that the metadata of the Zenodo deposit match the paper, in this case the title needs to be manually edited in the metadata. Could you make that tweak and let me know?
