### 2021-05-31 #111 @JesseBonanno
Hi @labarba I have just updated the metadata.

I changed the title to be "IndeterminateBeam: A Python package for solving 1D indeterminate beams" and changed the description to be an introductory segment from the paper.
